### 2023-11-09 #6042 @editorialbot
[PRE REVIEW]: PDOPT: A Python library for Probabilistic Design space exploration and OPTimisation.
**Submitting author:** @spinjet (<a href="http://orcid.org/0000-0002-3064-3387">Andrea Spinelli</a>)
**Repository:** https://github.com/spinjet/pdopt-code
**Branch with paper.md** (empty if default branch):
**Version:** v0.45
