### 2023-11-30 #6042 @kyleniemeyer
Hello @e-dub, could you edit this submission to JOSS?
