### 2023-11-30 #6042 @jbussemaker
> Hello @jbussemaker, could you edit this submission to JOSS?

Yes I would!
