### 2023-12-01 #6042 @zhaowei0566
> Hello @zhaowei0566, could you edit this submission to JOSS?

@kyleniemeyer  I am sorry I don't have sufficient time to review the manuscript and code this time.  I'm quite occupied these days as the end of the semester is approaching and I have some other responsibilities to wrap up. I'll make sure to take care of it next time.
