### 2023-12-01 #6110 @kyleniemeyer
👋🏼 @spinjet @e-dub, @jbussemaker this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
