### 2023-12-11 spinjet/pdopt-code#2 @jbussemaker
Improve code of some classes and functions
Suggestions to improve the DesignSpace object:
- [x] Currently the DesignSpace is defined from 2 csv file which are then internally converted to a nice OOP representation: I would suggest to make the OOP definition (ContinuousParameter, Objective, etc.) the main API for instantiating the DesignSpace, and add a helper function (a `@classmethod`) to create the DesignSpace from csv definitions
- [x] Add helper methods to the DesignSpace for saving to and loading from pickle

Probabilistic Exploration object:
- [x] Place less code in the `__init__` function: move the DoE and surrogate training code to a separate method
- [x] Add helper methods for saving to/loading from pickle

Extendable Model:
- [x] It is not clear from the base class what the `run` function actually does and what it takes as input / provides as output, only from reading code it seems like it takes design parameters as input and should return the outputs specified in the csv --> define this using type hints instead

Part of [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6110#issuecomment-1837524308)
