### 2023-12-11 spinjet/pdopt-code#3 @jbussemaker
FutureWarnings when running example 1
Environment: Python 12 env created with conda, dependencies installed with `pip install -r requirements.txt`

When I run `energy_management_experiment.py` in `example_1`, I continuously get the following warnings:
```
C:\...\pdopt-code\example\example_1\HE_Model.py:324: FutureWarning: Setting an item of incompatible dtype is deprecated and will raise in a future error of pandas. Value '[0.02935453 0.04485256 0.0603506  0.07584864 0.09134668 0.10684472
 0.12234276 0.1378408  0.15333884 0.16883688 0.18433491 0.19983295
 0.21533099 0.23082903 0.24632707 0.26182511 0.27732315 0.29282119
 0.30831923 0.32381726 0.3393153  0.35481334 0.37031138 0.38580942
 0.40130746 0.4168055  0.43230354 0.44780158 0.46329962 0.47879765]' has dtype incompatible with int64, please explicitly cast to a compatible dtype first.
  mission.loc[mission['tag'] == s, 'pDOH'] = tmp
C:\...\pdopt-code\example\example_1\HE_Model.py:324: FutureWarning: Setting an item of incompatible dtype is deprecated and will raise in a future error of pandas. Value '[0.24989405 0.24987103 0.24984757 0.24982366 0.24979927 0.24977438
 0.24974898 0.24972304 0.24969653 0.24966944 0.24964174 0.24961339
 0.24958437 0.24955465 0.24952418 0.24949293 0.24946086 0.24942793
 0.24939408 0.24935926 0.24932342 0.24928649 0.24924841 0.24920909
 0.24916847 0.24912644 0.24908291 0.24903776 0.24899086 0.24894209
 0.24889128]' has dtype incompatible with int64, please explicitly cast to a compatible dtype first.
  mission.loc[mission['tag'] == s, 'pDOH'] = tmp
```

The env has pandas 2.1.4 installed.

Part of [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6110#issuecomment-1837524308)
