### 2023-12-14 spinjet/pdopt-code#1 @kyleniemeyer
@jbussemaker regarding the author list: JOSS does not require that all paper authors have contributed source code, since people can contribute in other ways (direction, etc.)
