### 2023-12-14 #6110 @kyleniemeyer
Hi @spinjet, @jbussemaker has made some comments as issues over in the software repo. Can you respond/address them when you get a chance?
