### 2024-01-10 spinjet/pdopt-code#1 @spinjet
> I'm currently reviewing the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6110#issuecomment-1837524308), and I have some suggestions to improve code quality and usability. I did not install/run any code yet, I'll get to that soon.
>
> ### Software Engineering (Python)
> Important:
>
> * [ ]  Add a `.gitignore` file to remove all contents from the repository that should not be there, for example **pycache** folder, dist/build folders. You can go to https://www.toptal.com/developers/gitignore/ and enter Python, PyCharm, Windows, etc. to get a good default .gitignore file.
> * [ ]  Add tests, absolutely essential to ensure your software functions as intended, and keeps doing so as you add/change functionality over time. I recommend using pytest
>
>   * [ ]  Consider automatically running tests using Github actions CI/CD
> * [ ]  Use one way to declare dependencies: better to integrate this in the setup.py file (even better to use setup.cfg)
> * [ ]  Add a `pyproject.toml` file: this file tells Python dependency management tools what tool should be used to install the library (in PDOPT's case: setuptools). Some info: https://godatadriven.com/blog/a-practical-guide-to-setuptools-and-pyproject-toml/
> * [ ]  Add community guidelines for: contributing, reporting bugs, getting support
> * [ ]  Improve code styling, e.g. by adhereing to PEP8 or using a formatter like [black](https://github.com/psf/black)
> * [ ]  Remove unused code (`cpu_test.py`; a big chunk of `data.py`)
>
> Suggestions:
>
> * [ ]  Consider managing stable versions by creating tags and releases
> * [ ]  Consider publishing your code on PyPI
> * [ ]  In your current documentation file, Figure 1.2 seems to be missing some refs
> * [ ]  Consider hosting the documentation as a website, for example on [readthedocs.io](https://about.readthedocs.com/)
>
>   * [ ]  To create the documentation, consider using modern tools like [mkdocs](https://www.mkdocs.org/) and [Jupyter Notebooks](https://jupyter.org/) instead of a PDF file (I know LaTeX is quite common in research circles, but for software it's usually much nicer to navigate a website with search, and Jupyter Notebooks are great for explaining code as it allows mixing code, execution output and explanatory text in one interactive document)
>   * [ ]  You can use [mkdocstrings](https://mkdocstrings.github.io/) to automatically generate your API reference automatically (this also makes it easier to keep in sync with your actual code)
>
> ### Paper
> * [ ]  Add a comparison to other approaches/software ("state of the field")
>
>   * [ ]  All the paper references are own references of previous publications, I would also have expected some references to theory (e.g. the probabilistic model or set-based design)
>
> @kyleniemeyer I'd like to know your opinion about the paper authors: I can only find code contributions by Andrea Spinelli, however Timoleon Kipouros is also listed as an author

Dear @jbussemaker, I've started working on these and I've already introduced the .gitignore file.

Could you please elaborate on the dependencies...? I am a bit inexperienced on creating python packages, this is my first time.
I was using an external text file to load the dependencies, so it was easier while I was developing the software. Should I have them hardcoded in the setup.py file, or external? Do you perhaps have a tutorial or guideline link I could look into?
