### 2024-01-10 spinjet/pdopt-code#1 @jbussemaker
@spinjet yes sure! The link I provided for the point after the "dependencies" actually contains a bit about specifying dependencies in a `setup.cfg` file (https://godatadriven.com/blog/a-practical-guide-to-setuptools-and-pyproject-toml/).

Some other links I found through a quick google query:
- https://dev.to/2320sharon/making-sense-of-pyprojecttoml-setupcfg-and-setuppy-2o6m
- https://setuptools.pypa.io/en/latest/userguide/quickstart.html
  - In particular: https://setuptools.pypa.io/en/latest/userguide/declarative_config.html
