### 2024-01-10 spinjet/pdopt-code#1 @kyleniemeyer
Here is another good guide to making an installable package: https://packaging.python.org/en/latest/tutorials/packaging-projects/
