+++ requirements.txt
@@ -1,11 +1,0 @@
-numpy
-scipy
-matplotlib
-pandas
-plotly
-dash
-scikit-learn
-pymoo
-joblib
-tqdm
-psutil
