+++ setup.cfg
@@ -1,0 +1,31 @@
+[metadata]
+name = pdopt
+version = attr: data.__version__
+author = Andrea Spinelli
+author_email = andrea.spinelli@cranfield.ac.uk
+url = https://godatadriven.com/blog/a-practical-guide-to-setuptools-and-pyproject-toml
+description = Package for Probabilistic Design Space Exploration using Set-Based Design
+long_description = file: README.md
+long_description_content_type = text/markdown
+keywords = example, setuptools
+license = MIT License
+classifiers =
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3
+
+[options]
+packages = find:
+zip_safe = True
+include_package_data = True
+install_requires =
+	numpy
+	scipy
+	matplotlib
+	pandas
+	scikit-learn
+	pymoo
+	joblib
+	tqdm
+	
+[options.extras_require]
+viz = 	plotly, dash
