+++ setup.cfg
@@ -28,4 +28,4 @@ install_requires =
 	tqdm
 	
 [options.extras_require]
-viz = 	plotly, dash
+viz = 	plotly; dash
