### 2024-01-11 d1928ca
Update setup.cfg
