+++ src/pdopt/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Oct 15 11:28:15 2021
+
+@author: Andrea Spinelli
+"""
+
+__author__ = 'Andrea Spinelli'
+__copyright__ = 'Copyright 2021, all rights reserved'
+__status__ = 'Development'
+__version__ = 0.4.0
