### 2024-01-11 e41ab5b
Cleanup and README edits
Removed commented code in data.py, introduced community guidelines in README and changed installation guide.
