+++ src/pdopt/data.py
@@ -10,7 +10,7 @@ Module that contains all the data structures used within PDOPT.
 __author__ = 'Andrea Spinelli'
 __copyright__ = 'Copyright 2021, all rights reserved'
 __status__ = 'Development'
-__version__ = 0.4.0
+__version__ = '0.4.0'
 
 # Standard Library Imports
 from itertools import count, product
@@ -316,56 +316,12 @@ class DesignSet:
     
     def set_as_discarded(self):
         self.is_discarded = True
-    
-    
-    # def set_optimisation_problem(self, model, parameters, objectives, constraints):
-        
-    #     self.optimisation_problem = OptimisationProblem(model, 
-    #                                                     parameters, 
-    #                                                     objectives, 
-    #                                                     constraints, 
-    #                                                     self.parameter_levels_list)#,
-    #                                                     #runner=pool.starmap,
-    #                                                     #func_eval=starmap_parallelized_eval)
+
     
     def set_optimisation_problem(self, opt_problem):
         
         self.optimisation_problem = opt_problem
-                                                        
-    # def set_surrogate_optimisation_problem    (self, model, parameters, objectives, constraints):
-        
-    #     self.surrogate_optimisation_problem = SurrogateOptimisationProblem(model, 
-    #                                                     parameters, 
-    #                                                     objectives, 
-    #                                                     constraints, 
-    #                                                     self.parameter_levels_list)
-    
-    # def set_robust_optimisation_problem(self, model, surrogates, parameters, objectives, constraints,
-    #                                     P_g, k_sigma, pool, decoupled_sigma=False):
-        
-    #     if decoupled_sigma:
-    #         self.rbo_problem =  RobustOptimisationProblem_Decoupled(model, 
-    #                                                         surrogates,   
-    #                                                         parameters, 
-    #                                                         objectives, 
-    #                                                         constraints, 
-    #                                                         self.parameter_levels_list,
-    #                                                         runner=pool.starmap,
-    #                                                         func_eval=starmap_parallelized_eval,
-    #                                                         P_constraints=P_g)
-        
-    #     else:
-    #         self.rbo_problem = RobustOptimisationProblem(model, 
-    #                                                         surrogates,   
-    #                                                         parameters, 
-    #                                                         objectives, 
-    #                                                         constraints, 
-    #                                                         self.parameter_levels_list,
-    #                                                         runner=pool.starmap,
-    #                                                         func_eval=starmap_parallelized_eval,
-    #                                                         P_constraints=P_g,
-    #                                                         k_sigma=3)
-        
+
 
     def sample(self, n_samples, parameters_list):
         # Sample a n_amount within the design space using Latin Hypercube
@@ -532,779 +488,6 @@ class DesignSet:
     def get_robust_optimum(self):
         return self.rbo_results
     
-
-# class SurrogateOptimisationProblem(Problem):
-#     def __init__(self, model, parameters, objectives, constraints, set_levels,
-#                  **kwargs):
-        
-#         self.model = model #store reference to model
-#         self.var = parameters
-#         self.obj = objectives
[…]
