+++ setup.cfg
@@ -1,13 +1,13 @@
 [metadata]
 name = pdopt
-version = attr: pdopt.data.__version__
+version = attr: pdopt.__version__
 author = Andrea Spinelli
 author_email = andrea.spinelli@cranfield.ac.uk
-url = https://godatadriven.com/blog/a-practical-guide-to-setuptools-and-pyproject-toml
+url = https://github.com/spinjet/pdopt-code
 description = Package for Probabilistic Design Space Exploration using Set-Based Design
 long_description = file: README.md
 long_description_content_type = text/markdown
-keywords = example, setuptools
+keywords = set-based design, design space exploration, optimization, engineering design
 license = MIT License
 classifiers =
     License :: OSI Approved :: MIT License
@@ -15,6 +15,8 @@ classifiers =
 
 [options]
 packages = find:
+package_dir=
+    =src
 zip_safe = True
 include_package_data = True
 install_requires =
@@ -26,6 +28,9 @@ install_requires =
 	pymoo
 	joblib
 	tqdm
+
+[options.packages.find]
+where=src
 	
 [options.extras_require]
 viz = 	plotly; dash
