### 2024-01-11 34a2d2b
PEP8 Compliance
Python scrips modified to be PEP8 compliant
