+++ src/pdopt/visualisation.py
@@ -5,10 +5,6 @@ Created on Mon Oct 25 09:55:50 2021
 @author: s345001
 """
 
-__author__ = 'Andrea Spinelli'
-__copyright__ = 'Copyright 2021, all rights reserved'
-__status__ = 'Development'
-
 # Standard Library Imports
 import base64
 import io
@@ -22,7 +18,6 @@ import pandas as pd
 
 import dash
 from dash import dcc, html
-#import dash_html_components as html
 from dash.dependencies import Input, Output, State
 
 import plotly.graph_objects as go
@@ -33,71 +28,73 @@ import numpy as np
 
 import plotly.io as pio
 
-pio.renderers.default='browser'
+pio.renderers.default = 'browser'
 
 # Local imports
-#from pdopt.data import DesignSpace, Model
 
 # Module Constants
+__author__ = 'Andrea Spinelli'
+__copyright__ = 'Copyright 2021, all rights reserved'
+__status__ = 'Development'
+
 pd.options.mode.chained_assignment = None
 
-NONE_LABELS = [{'label' : 'None', 'value' : 'none'}]
+NONE_LABELS = [{'label': 'None', 'value': 'none'}]
 
 LOAD_MODULE = html.Div(children=[
-                dcc.Upload(id='upload-data', 
-                          children=[
-                              html.A('Load .csv file')],
-                          style={
-                                'width': 'fit-content',
-                                'height': '60px',
-                                'lineHeight': '60px',
-                                'borderWidth': '1px',
-                                'borderStyle': 'dashed',
-                                'borderRadius': '5px',
-                                'textAlign': 'center',
+    dcc.Upload(id='upload-data',
+               children=[
+                   html.A('Load .csv file')],
+               style={
+                   'width': 'fit-content',
+                   'height': '60px',
+                   'lineHeight': '60px',
+                   'borderWidth': '1px',
+                   'borderStyle': 'dashed',
+                   'borderRadius': '5px',
+                   'textAlign': 'center',
                                 'margin': '10px',
                                 'float': 'left',
                                 'display': 'inline-block',
                                 'padding': '1px 10px',
-                            },)
-                ])
-
+               },)
+])
 
 
 # Scatter Sizes
 SCATTER_SIZES = ['45vw', '45vw', '31vw', '31vw', '31vw']
 
 # Color Schemes
-SCATTER_COLOR  = 'rainbow'
-PARCOORD_COLOR  = 'turbo'
+SCATTER_COLOR = 'rainbow'
+PARCOORD_COLOR = 'turbo'
 PARCOORD_COLOR_LIST = [
-                'rgb(48,18,59)',
-                'rgb(65,69,171)',
-                'rgb(70,117,237)',
-                'rgb(57,162,252)',
-                'rgb(27,207,212)',
-                'rgb(36,236,166)',
-                'rgb(97,252,108)',
-                'rgb(164,252,59)',
-                'rgb(209,232,52)',
-                'rgb(243,198,58)',
-                'rgb(254,155,45)',
-                'rgb(243,99,21)',
-                'rgb(217,56,6)',
-                'rgb(177,25,1)',
-                'rgb(122,4,2)',]
+    'rgb(48,18,59)',
+    'rgb(65,69,171)',
+    'rgb(70,117,237)',
+    'rgb(57,162,252)',
+    'rgb(27,207,212)',
+    'rgb(36,236,166)',
+    'rgb(97,252,108)',
+    'rgb(164,252,59)',
+    'rgb(209,232,52)',
[…]
