+++ src/pdopt/data.py
@@ -7,18 +7,13 @@ Module that contains all the data structures used within PDOPT.
 @author: Andrea Spinelli
 """
 
-__author__ = 'Andrea Spinelli'
-__copyright__ = 'Copyright 2021, all rights reserved'
-__status__ = 'Development'
-__version__ = '0.4.0'
-
 # Standard Library Imports
 from itertools import count, product
 from abc import ABC, abstractmethod
 from time import time
 
 # Third-party imports
-import numpy  as np
+import numpy as np
 import pandas as pd
 
 from scipy.stats.qmc import LatinHypercube, Sobol
@@ -27,6 +22,10 @@ from scipy.stats import norm, triang, uniform
 # Local imports
 
 # Module Constants
+__author__ = 'Andrea Spinelli'
+__copyright__ = 'Copyright 2021, all rights reserved'
+__status__ = 'Development'
+__version__ = '0.4.0'
 
 # Module Functions
 
@@ -34,154 +33,153 @@ from scipy.stats import norm, triang, uniform
 # Abstract Classes
 class Base:
     _ids = count(0)
-    
+
     def __init__(self):
         pass
 
 
 class Parameter:
     _ids = count(0)
-    
-    
+
     def get_id(self):
         return self.id
-    
-    
+
     def get_name(self):
         return self.name
-    
-    
+
     def __init__(self, name):
-        self.id   = next(self._ids)
+        self.id = next(self._ids)
         self.name = name
-        
-        
+
     def __repr__(self):
         pass
 
+
 class Response:
     _ids = count(0)
-    
+
     def get_id(self):
         return self.id
-    
-    
+
     def get_name(self):
         return self.name
-    
-    
+
     def __init__(self, name, operand, value, p_satisfaction):
-        self.id             = next(self._ids)
-        self.name           = name
-        self.operand        = operand
-        self.value          = value
-        #Min P to satisfy the response, for the sample
+        self.id = next(self._ids)
+        self.name = name
+        self.operand = operand
+        self.value = value
+        # Min P to satisfy the response, for the sample
         self.p_satisfaction = p_satisfaction
+
 
 class ExtendableModel:
     # Model Object that can be extended and used by the library
     def __init__(self):
         pass
-    
+
     def run(self):
         pass
-        
+
 # Concrete Classes
+
+
 class ContinousParameter(Parameter):
-    
+
     def get_bounds(self):
         return self.lb, self.ub
-    
-    
+
     def get_level_bounds(self, level):
         assert level < self.n_levels, "Selected level is above total number of levels"
-        return self.ranges[level], self.ranges[level+1]
-    
-    
+        return self.ranges[level], self.ranges[level + 1]
+
     def __init__(self, name, lb, ub, n_levels, uq_dist, uq_var_l, uq_var_u):
         super().__init__(name)
-        
+
         assert lb < ub, "Lower Bound value higher than Upper Bound value"
         self.lb, self.ub = lb, ub
         self.n_levels = n_levels
-        
-        self.ranges  = tuple(np.linspace(lb, ub, n_levels+1))
+
+        self.ranges = tuple(np.linspace(lb, ub, n_levels + 1))
         self.uq_dist = uq_dist
         self.uq_var_l = uq_var_l
         self.uq_var_u = uq_var_u
-        
+
     def __repr__(self):
         # Print information on the Parameter
-        
-        s1 = 'id:{:d} - Continous Parameter "{}"\n{:d} Levels, '.format(self.id, 
-                                                                        self.name, 
+
+        s1 = 'id:{:d} - Continous Parameter "{}"\n{:d} Levels, '.format(self.id,
+                                                                        self.name,
[…]
