+++ src/pdopt/__init__.py
@@ -4,4 +4,3 @@ Created on Fri Oct 15 11:28:15 2021
 
 @author: Andrea Spinelli
 """
-

[…]
