### 2024-01-11 spinjet/pdopt-code#1 @spinjet
Progress so far:

### Software Engineering (Python)

Important:
- [x] Add a `.gitignore` file to remove all contents from the repository that should not be there, for example __pycache__ folder, dist/build folders. You can go to https://www.toptal.com/developers/gitignore/ and enter Python, PyCharm, Windows, etc. to get a good default .gitignore file.
- [ ] Add tests, absolutely essential to ensure your software functions as intended, and keeps doing so as you add/change functionality over time. I recommend using pytest
  - [ ] Consider automatically running tests using Github actions CI/CD
- [x] Use one way to declare dependencies: better to integrate this in the setup.py file (even better to use setup.cfg)
- [x] Add a `pyproject.toml` file: this file tells Python dependency management tools what tool should be used to install the library (in PDOPT's case: setuptools). Some info: https://godatadriven.com/blog/a-practical-guide-to-setuptools-and-pyproject-toml/
- [x] Add community guidelines for: contributing, reporting bugs, getting support
- [x] Improve code styling, e.g. by adhereing to PEP8 or using a formatter like [black](https://github.com/psf/black)
- [x] Remove unused code (`cpu_test.py`; a big chunk of `data.py`)

Suggestions:
- [ ] Consider managing stable versions by creating tags and releases
- [ ] Consider publishing your code on PyPI
- [ ] In your current documentation file, Figure 1.2 seems to be missing some refs
- [ ] Consider hosting the documentation as a website, for example on [readthedocs.io](https://about.readthedocs.com/)
  - [ ] To create the documentation, consider using modern tools like [mkdocs](https://www.mkdocs.org/) and [Jupyter Notebooks](https://jupyter.org/) instead of a PDF file (I know LaTeX is quite common in research circles, but for software it's usually much nicer to navigate a website with search, and Jupyter Notebooks are great for explaining code as it allows mixing code, execution output and explanatory text in one interactive document)
  - [ ] You can use [mkdocstrings](https://mkdocstrings.github.io/) to automatically generate your API reference automatically (this also makes it easier to keep in sync with your actual code)

- [x] Add a comparison to other approaches/software ("state of the field")
  - [x] All the paper references are own references of previous publications, I would also have expected some references to theory (e.g. the probabilistic model or set-based design)

@kyleniemeyer, @jbussemaker
Thank you so far for your help, I've adopted the new .toml approach with a setuptools .cfg file. I'll try to have it uploaded on PyPI for easy install through `pip`, albeit I may not be able to maintain the code 24/7.

I'm going to be addressing the tests soon, I'll try to add a few using `pytest`.
I'll also address the paper and extend it with some background referencing directly some of the building blocks that aren't directly developed by me (Set-based design being one). I'm not sure for other approaches, I reckon I could mention some design space exploration approaches or general use of optimisation as a discovery tool.
