### 2024-01-11 spinjet/pdopt-code#1 @kyleniemeyer
Hi @spinjet - good progress! I might suggest that for simplicity, you just include your package's dependencies directly in the `pyproject.toml` file, rather than a separate file: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies
