### 2024-01-11 spinjet/pdopt-code#1 @jbussemaker
And I just wanted to note that publishing on PyPI is also possible by CI/CD workflow, to automate most of it:

https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/
