### 2024-01-12 69c1844
Stub of testing using pytest
