+++ tests/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Oct 15 11:28:15 2021
+
+@author: Andrea Spinelli
+"""
