+++ tests/test_data.py
@@ -7,15 +7,190 @@ Created on Fri Jan 12 11:45:20 2024
 
 import pytest
 import pdopt.data as data
+import numpy as np
+import pandas as pd
+import os
 
-
+# Test the continous parameter data structure works
 @pytest.fixture
 def my_continous_parameter():
-    return data.data.ContinousParameter('test_continous', 
-                                        0, 1, 5, 
-                                        uq_dist, uq_var_l, uq_var_u)
+    return data.ContinousParameter('test_continous', 
+                                        0, 1, 4, 
+                                        'uniform', 0.1, 0.1)
 
 
+def test_continous_parameter_repr(my_continous_parameter):
+    test_string = f'id:{my_continous_parameter.id} - Continous Parameter "{my_continous_parameter.name}"\n{my_continous_parameter.n_levels} Levels, Bounds: (0.000, 0.250, 0.500, 0.750, 1.000)\n'
+    assert str(my_continous_parameter) == test_string
 
-        
-        
+
+def test_continous_parameter_bounds(my_continous_parameter):
+    assert my_continous_parameter.get_bounds() == (0, 1)
+    
+@pytest.mark.parametrize("level, expected_output", [(0, (0.0, 0.25)), (1, (0.25, 0.50)), 
+                                   (2, (0.50, 0.75)), (3, (0.75, 1.0))])
+def test_continous_parameter_get_level_bounds(my_continous_parameter, level, expected_output):
+    assert my_continous_parameter.get_level_bounds(level) == expected_output
+    
+
+# Test the continous parameter percentile works for the supported distributions
+@pytest.mark.parametrize("distribution, expected_output", 
+                         [('uniform', np.array([0.475, 0.5, 0.525])),
+                          ('triang', np.array([0.48535534, 0.5, 0.51464466])),
+                          ('norm', np.array([0.48313776, 0.5, 0.51686224])),
+                          (None, np.array([0.5, 0.5, 0.5]))])
+def test_continous_parameter_distributions_ppf(distribution, expected_output):
+    my_parameter = data.ContinousParameter('test_continous', 
+                                        0, 1, 4, 
+                                        distribution, 0.1, 0.1)
+    
+    assert (my_parameter.ppf([0.25, 0.5, 0.75], 0.5).round(3) == expected_output.round(3)).all()
+
+
+# Test the discrete parameter data structure works
+@pytest.fixture
+def my_discrete_parameter():
+    return data.DiscreteParameter('test_discrete', 5)
+
+def test_discrete_parameter_repr(my_discrete_parameter):
+    assert str(my_discrete_parameter) == f'id:{my_discrete_parameter.id} - Discrete Parameter "{my_discrete_parameter.name}"\n{my_discrete_parameter.n_levels} Levels\n'
+    
+def test_discrete_parameter_get_n_levels(my_discrete_parameter):
+    assert my_discrete_parameter.get_n_levels() == 5
+    
+    
+# Test the objective object
+def test_objective_repr():
+    my_objective = data.Objective('test_objective', 'min', min_requirement=1, p_sat=0.3)
+    assert str(my_objective) == f"Objective {my_objective.name} : {my_objective.operand}: Requirement: {my_objective.get_requirement()} : P_sat = {my_objective.p_satisfaction}\n"
+
+@pytest.mark.parametrize("operand, has_requirement, expected_output", 
+                         [('min', None, None), ('max', None, None),
+                          ('min', 1, ('lt', 1)), ('max', 1, ('gt', 1)),])
[…]
