+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pdopt
-version = attr: pdopt.__version__
+version = attr: pdopt.data.__version__
 author = Andrea Spinelli
 author_email = andrea.spinelli@cranfield.ac.uk
 url = https://github.com/spinjet/pdopt-code
