+++ .github/workflows/pytest.yml
@@ -1,0 +1,41 @@
+ name: pytest
+ on: 
+    push:
+    pull_request:
+    
+jobs:
+    run-tests:
+        strategy:
+            fail-fast: false
+            matrix:
+                os: [ubuntu-latest, macos-latest, windows-latest]
+                python-version: 
+                    - "3.7"
+                    - "3.8"
+                    - "3.9"
+                    - "3.10"
+                    - "3.11"
+                    - "3.12"
+        name: Test
+        runs-on: ${{ matrix.os }}
+        
+        
+        steps:
+            - name: Checkout code
+              uses: actions/checkout@v3
+            
+            - name: Setup Python 
+              uses: actions/setup-python@v4 
+              with:
+                python-version: ${{ matrix.python-version }}
+                
+            - name: Install dependencies
+              run: |
+                python -m pip install pytest .
+                python -m pip install -e .
+
+            - name: Run test 
+              run: |
+                cd .\tests
+                pytest 
+            
