+++ .github/workflows/pytest.yml
@@ -1,7 +1,5 @@
 name: pytest
-on: 
-push:
-  pull_request:
+on: [push]
     
 jobs:
   test:
@@ -9,13 +7,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: 
-          - "3.7"
-          - "3.8"
-          - "3.9"
-          - "3.10"
-          - "3.11"
-          - "3.12"
+        python-version: ["pypy3.9", "pypy3.10", "3.9", "3.10", "3.11", "3.12"]
         name: test
         runs-on: ${{ matrix.os }}
         
