+++ .github/workflows/pytest.yml
@@ -8,26 +8,24 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ["pypy3.9", "pypy3.10", "3.9", "3.10", "3.11", "3.12"]
-        name: test
-        runs-on: ${{ matrix.os }}
-        
-        
-        steps:
-          - name: Checkout code
-            uses: actions/checkout@v3
-          
-          - name: Setup Python 
-            uses: actions/setup-python@v4 
-            with:
-              python-version: ${{ matrix.python-version }}
-              
-          - name: Install dependencies
-            run: |
-              python -m pip install pytest .
-              python -m pip install -e .
-          
-          - name: Run test 
-            run: |
-              cd .\tests
-              pytest 
+    name: test
+    runs-on: ${{ matrix.os }}
+    steps:
+     - name: Checkout code
+       uses: actions/checkout@v3
+     
+     - name: Setup Python 
+       uses: actions/setup-python@v4 
+       with:
+         python-version: ${{ matrix.python-version }}
+         
+     - name: Install dependencies
+       run: |
+         python -m pip install pytest .
+         python -m pip install -e .
+     
+     - name: Run test 
+       run: |
+         cd .\tests
+         pytest 
             
