### 2024-01-17 da68608
Update setup.cfg
