+++ setup.cfg
@@ -25,7 +25,7 @@ install_requires =
 	matplotlib
 	pandas
 	scikit-learn
-	pymoo>="0.6.0"
+	pymoo >= 0.6.0
 	joblib
 	tqdm
 
