### 2024-01-17 e6ecb42
Remove pypy testing as it is not supported.
