+++ tests/test_exploration.py
@@ -1,0 +1,24 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Jan 17 12:35:33 2024
+
+@author: Andrea Spinelli
+"""
+
+import pytest
+import pdopt.exploration as exploration
+import numpy as np
+import pandas as pd
+
+## Test generate input samples
+
+
+
+## Test training data generation
+
+## Test testing data generation
+
+## Test surrogate response
+
+## Test probabilistic exploration
+
