### 2024-01-17 5fe9bb6
Partially written test for exploration.py
