### 2024-01-19 4a852d2
Implemented automated tests
Fully implemented automated tests to check the integrity of the package.
