+++ .github/workflows/pytest.yml
@@ -27,5 +27,11 @@ jobs:
      - name: Run test 
        run: |
          cd ./tests
-         pytest 
-            
+         pytest -vv > test-result-${{ matrix.os }}-${{ matrix.python-version}}.txt
+     
+     - name: Upload results
+       uses: actions/upload-artifact@v3
+       with:
+         name: pytest-results-${{ matrix.os }}-${{ matrix.python-version}}
+         path: test-results/test-result-${{ matrix.os }}-${{ matrix.python-version}}.txt
+       if: ${{ always() }}

[…]
