### 2024-01-19 fdbca7c
Update pytest.yml
