### 2024-01-19 c7c1483
Update pytest.yml
Add back test results
