### 2024-01-19 649fe44
Update test_optimisation.py
