### 2024-01-22 spinjet/pdopt-code#1 @spinjet
Dear @kyleniemeyer @jbussemaker
Thank you again for your help.

I've been writing tests and integrating them with pytest and GitHub actions, I am close to completion to this task.
I need to fix the testing of the optimisation module as it seems there's some randomness I cannot control (the module relies on stochastic optimisation approaches).

Hopefully this should be done by the week along with the CI for publishing!
You should be able to see the progress in the repository commits.
