+++ tests/test_optimisation.py
@@ -143,12 +143,12 @@ def test_NNSurrogate_train_model_predict(my_NNSurrogate):
     my_NNSurrogate.train_model()
     
     #Test the prediction    
-    expected_output = {'F': np.array([[22.39651836],
-                            [19.88252189],
-                            [17.81039263]]),
-                     'G': np.array([[-22.78313439,   0.97456643],
-                                    [-20.19785334,   0.85587712],
-                                    [-18.05448147,   0.70555098]])}
+    expected_output = {'F': np.array([[22.396518358255157],
+                                       [19.882521893338144],
+                                       [17.810392630436965]]),
+                     'G': np.array([[-22.78313438982879  ,   0.9745664293583953],
+                                    [-20.197853335738785 ,   0.8558771222671897],
+                                    [-18.054481467775297 ,   0.7055509759847869]])}
     
     x_in     = np.array([[0.25, 0],
                          [0.5,  0],
@@ -157,8 +157,8 @@ def test_NNSurrogate_train_model_predict(my_NNSurrogate):
     
     my_NNSurrogate._evaluate(x_in, out_dict)
     
-    assert (out_dict['F'].round(3) == expected_output['F'].round(3)).all()
-    assert (out_dict['G'].round(3) == expected_output['G'].round(3)).all()
+    assert (out_dict['F'].round(6) == expected_output['F'].round(6)).all()
+    assert (out_dict['G'].round(6) == expected_output['G'].round(6)).all()
 
 
 def test_NSSurrogate_reconstruct_recover_pts(my_NNSurrogate):
@@ -169,9 +169,9 @@ def test_NSSurrogate_reconstruct_recover_pts(my_NNSurrogate):
                          [0.5,  0],
                          [0.75, 0]])
 
-    expected_output = np.array([[ 0.25  ,  0.     , 22.5625 , 22.5625 , 10.96875],
-                               [ 0.5    ,  0.     , 20.25   , 20.25   , 10.875  ],
-                               [ 0.75   ,  0.     , 18.0625 , 18.0625 , 10.71875]])
+    expected_output = np.array([[ 0.25000,  0.00000, 22.56250, 22.56250, 10.96875],
+                               [ 0.50000,  0.00000, 20.25000, 20.25000, 10.87500],
+                               [ 0.75000,  0.00000, 18.06250, 18.06250, 10.71875]])
                             
     out = my_NNSurrogate.recover_pts(x_in)
     
@@ -297,12 +297,12 @@ def test_KrigingSurrogate_train_model_predict(my_KrigingSurrogate):
     my_KrigingSurrogate.train_model()
     
     #Test the prediction    
-    expected_output = {'F': np.array([[22.56252049],
-                            [20.25003194],
-                            [18.06250782]]),
-                     'G': np.array([[-22.56252049,   0.96874912],
-                            [-20.25003194,   0.87499755],
-                            [-18.06250782,   0.718747  ]])}
+    expected_output = {'F': np.array([[22.562512195037218],
+                            [20.250022440354236],
+                            [18.062502637980753]]),
+                     'G': np.array([[-22.5625121950372183,   0.9687476713857812],
+                                    [-20.2500224403542362,   0.8749941943046551],
+                                    [-18.0625026379807530,   0.7187488119140824]])}
     
     x_in     = np.array([[0.25, 0],
                          [0.5,  0],
@@ -311,8 +311,8 @@ def test_KrigingSurrogate_train_model_predict(my_KrigingSurrogate):
     
     my_KrigingSurrogate._evaluate(x_in, out_dict)
     
-    assert (out_dict['F'].round(3) == expected_output['F'].round(3)).all()
-    assert (out_dict['G'].round(3) == expected_output['G'].round(3)).all()
[…]
