+++ tests/test_optimisation.py
@@ -157,8 +157,11 @@ def test_NNSurrogate_train_model_predict(my_NNSurrogate):
     
     my_NNSurrogate._evaluate(x_in, out_dict)
     
-    assert (out_dict['F'].round(6) == expected_output['F'].round(6)).all()
-    assert (out_dict['G'].round(6) == expected_output['G'].round(6)).all()
+    # There is some small numerical error that I cannot control caused by the
+    # type of machine used for testing by GitHub. Hence assert is written 
+    # to assert this precision
+    assert (abs(out_dict['F'] - expected_output['F']) < 1e-4).all()
+    assert (abs(out_dict['G'] - expected_output['G']) < 1e-4).all()
 
 
 def test_NSSurrogate_reconstruct_recover_pts(my_NNSurrogate):
@@ -311,8 +314,11 @@ def test_KrigingSurrogate_train_model_predict(my_KrigingSurrogate):
     
     my_KrigingSurrogate._evaluate(x_in, out_dict)
     
-    assert (out_dict['F'].round(6) == expected_output['F'].round(6)).all()
-    assert (out_dict['G'].round(6) == expected_output['G'].round(6)).all()
+    # There is some small numerical error that I cannot control caused by the
+    # type of machine used for testing by GitHub. Hence assert is written 
+    # to assert this precision
+    assert (abs(out_dict['F'] - expected_output['F']) < 1e-4).all()
+    assert (abs(out_dict['G'] - expected_output['G']) < 1e-4).all()
 
 @pytest.mark.filterwarnings("ignore: lbfgs failed")
 def test_KrigingSurrogate_reconstruct_recover_pts(my_KrigingSurrogate):
@@ -385,8 +391,10 @@ def test_DirectOpt(my_Model, my_DirectOpt):
                                    [ 4.0000000000000000e+00,  5.0000000000003437e+00,
                                      0.0000000000000000e+00,  1.1814690891404572e-25,
                                      1.1814690891404572e-25, -1.5000000000017195e+00]]))])
-def test_Optimisation_run(my_Model, my_DesignSpace, 
+def test_Optimisation_run(my_Model, 
                     use_surrogate, use_nn, gp_kern, expected_output):
+    
+    my_DesignSpace = data.DesignSpace('test_input.csv', 'test_response.csv')
     my_Exploration = exploration.ProbabilisticExploration(my_DesignSpace, 
                                                 my_Model,
                                                 debug=True)
@@ -400,6 +408,10 @@ def test_Optimisation_run(my_Model, my_DesignSpace,
                                                 debug=True)
     
     my_Optimisation.run('.')
-    assert (my_DesignSpace.get_optimum_results().to_numpy().round(6) == expected_output.round(6)).all()
+    
+    # There is some small numerical error that I cannot control caused by the
+    # type of machine used for testing by GitHub. Hence assert is written 
+    # to assert this precision
+    assert (abs(my_DesignSpace.get_optimum_results().to_numpy() - expected_output) < 1e-4).all()
     
     
