### 2024-01-22 ed63fc5
Updated tests
Fixed the number of optimisation iterations to minimise any variability.
