+++ tests/test_optimisation.py
@@ -366,35 +366,35 @@ def test_DirectOpt(my_Model, my_DirectOpt):
 @pytest.mark.filterwarnings("ignore: invalid")
 @pytest.mark.parametrize('use_surrogate, use_nn, gp_kern, expected_output',
                           [(True, False, 'matern', 
-                            np.array([[ 2.0000000000000000e+00,  4.9789455535613270e+00,
-                                     0.0000000000000000e+00,  4.4328971483895217e-04,
-                                     4.4328971483895217e-04, -1.3949494126640545e+00],
-                                   [ 4.0000000000000000e+00,  5.0150069987293389e+00,
-                                     0.0000000000000000e+00,  2.2521001086237818e-04,
-                                     2.2521001086237818e-04, -1.5751475986521246e+00]])),
+                            np.array([[ 2.0000000000000000e+00,  4.9776166868405136e+00,
+                                     0.0000000000000000e+00,  5.0101270799563630e-04,
+                                     5.0101270799563630e-04, -1.3883339405565653e+00],
+                                   [ 4.0000000000000000e+00,  5.0153314976974972e+00,
+                                     0.0000000000000000e+00,  2.3505482164836190e-04,
+                                     2.3505482164836190e-04, -1.5767750158983098e+00]])),
                           (True, False, 'rbf', 
-                           np.array([[ 2.0000000000000000e+00,  4.9938241170193960e+00,
-                                    0.0000000000000000e+00,  3.8141530590114250e-05,
-                                    3.8141530590114250e-05, -1.4691396558622749e+00],
-                                  [ 4.0000000000000000e+00,  5.0051232487877719e+00,
-                                    0.0000000000000000e+00,  2.6247678141406529e-05,
-                                    2.6247678141406529e-05, -1.5256293677779311e+00]])),
+                           np.array([[ 2.0000000000000000e+00,  4.9930717283301140e+00,
+                                    0.0000000000000000e+00,  4.8000948331744473e-05,
+                                    4.8000948331744473e-05, -1.4653826421247356e+00],
+                                  [ 4.0000000000000000e+00,  5.0052569547100640e+00,
+                                    0.0000000000000000e+00,  2.7635572823664532e-05,
+                                    2.7635572823664532e-05, -1.5262985913367313e+00]])),
                           (True, True,  None, 
-                            np.array([[ 2.0000000000000000,  4.8433069697945648,  0.0000000000000000,
-                                     0.0245527057149614,  0.0245527057149614, -0.7288112018303039],
-                                   [ 4.0000000000000000,  5.1524178628201662,  0.0000000000000000,
-                                     0.0232312049066670,  0.0232312049066670, -2.2737049165541645]])),
+                            np.array([[ 2.0000000000000000,  4.8424522210498493,  0.0000000000000000,
+                                     0.0248213026521256,  0.0248213026521256, -0.7246717565753098],
+                                   [ 4.0000000000000000,  5.1613830538504022,  0.0000000000000000,
+                                     0.0260444900700818,  0.0260444900700818, -2.3199375142870515]])),
                           (False, False, None, 
-                            np.array([[ 2.0000000000000000e+00,  4.9999999999981739e+00,
-                                     0.0000000000000000e+00,  3.3346223314523270e-24,
-                                     3.3346223314523270e-24, -1.4999999999908695e+00],
-                                   [ 4.0000000000000000e+00,  5.0000000000003437e+00,
-                                     0.0000000000000000e+00,  1.1814690891404572e-25,
-                                     1.1814690891404572e-25, -1.5000000000017195e+00]]))])
+                            np.array([[ 2.0000000000000000e+00,  4.9997891005059012e+00,
+                                     0.0000000000000000e+00,  4.4478596611131480e-08,
[…]
