+++ .github/workflows/pytest.yml
@@ -27,11 +27,4 @@ jobs:
      - name: Run test 
        run: |
          cd ./tests
-         pytest -vv | tee ./test-results/test-result-${{ matrix.os }}-${{ matrix.python-version}}.log
-     
-     - name: Upload results
-       uses: actions/upload-artifact@v3
-       with:
-         name: pytest-results-${{ matrix.os }}-${{ matrix.python-version}}
-         path: tests/test-reulsts/test-result-${{ matrix.os }}-${{ matrix.python-version}}.log
-       if: ${{ always() }}
+         pytest -vv 
