### 2024-01-22 #6110 @kyleniemeyer
For awareness, some discussion around updates to the paper is happening at https://github.com/spinjet/pdopt-code/issues/1
