### 2024-01-23 spinjet/pdopt-code#2 @spinjet
Dear @jbussemaker

I'm going through these suggestions, so I have a question.

Regarding the DesignSpace object, the original idea was to use the .csv files as main input because it would allow the user to setup multiple cases to be ran in parallel from the terminal, hence changing just those input files.
I suppose instantiation would work in code for small problems. However, I am not sure how to express it in a OOP way... I reckon I'd need a setter function for each parameter?
