### 2024-01-23 spinjet/pdopt-code#2 @jbussemaker
Good question @spinjet

So what I meant was: why is it needed to pass in a csv files? Of course, in some cases it can be a convenient interface, but for the functioning of the program it is not *really* needed. Actually, already during instantiation you're reading the csv files and converting them to another format, namely a set of Parameter, Objective and Constraint objects.

My suggestion would be to change the main input to the `__init__` function to be these lists of parameters, objective and constraints directly, so that the user can conveniently use the DesignSpace object purely from within Python. Specifying function/class inputs/outputs as objects also makes the code self-documenting, as it becomes clear exactly what data to provide and in which format. It also would make unit testing easier, since it's all just Python.

If you then still want to keep the possibility for using the csv files as an input (because I guess it fits to your workflow, and would need less modification of the documentation), then you can do so using a class factory function:
```python
class DesignSpace:

    def __init__(params, obj, con, ...):
        pass  # Initialize the class using params, obj, con objects

    @classmethod
    def from_csv(cls, csv_parameters, csv_responses):
        # Read the CSV files and convert them to params, obj, con objects
        params = ...
        obj = ...
        con = ...

        # Then create the class
        return cls(params, obj, con)

# Usage example:
design_space = DesignSpace.from_csv(...)
```

Like this you could also add other factory functions if the need for that arises in the future (for example to load from a pickle-serialized DesignSpace or something like that).

These principles apply to several other objects as well if I remember correctly.
