### 2024-01-23 30734a5
Improved ProbabilisticExploration class
Moved code inside the __init__() method to specific private methods for easier reading and debugging. Updated the relative test code to check the functionality of these private methods.
