### 2024-01-23 d2bf7d2
Added type hinting to the ExtendableModel class.
