### 2024-01-23 spinjet/pdopt-code#1 @spinjet
As mentioned in https://github.com/openjournals/joss-reviews/issues/6110#issuecomment-1906266871
Automated testing with CI is now fully implemented.
