### 2024-01-23 e9f9f21
Added from_csv classmethod
Standardised the creation interface of the DesignSpace object to just python objects and moved the loading from csv to a helper method.
