### 2024-01-23 0f1b6ae
Added pickling interface to DesignSpace object
