### 2024-01-23 spinjet/pdopt-code#2 @spinjet
Added the ability to load DesignSpace from csv and pickle through classmethods, plus a saving method.
I have created a branch with these changes:
https://github.com/spinjet/pdopt-code/tree/dev-joss-review

I plan to also add a similar interface for saving the data in .csv files through Pandas, just to polish it.
