### 2024-01-23 spinjet/pdopt-code#4 @e-dub
Review suggestions, part II
So far impressive stuff!  I am starting a new issue for my comments as part of the: [JOSS review](https://github.com/openjournals/joss-reviews/issues/6110#issuecomment-1906815430).  Work in progress...

...

- [x] `Contemporary...`
I do not know if only contemporary systems are characterized in this way. What is different now is that we are now trying better control/master/handle/design with this complexity, no?
- [x] You are framing the context within the realm of uncertainty in the summary.  Do I understand correctly that the surrogate model is able to capture the design space and its uncertainty?  Or does the surrogate model work in a probabilistic way?  Can you please add a sentence to make this clearer?

- [x] `The design of such systems entails high uncertainty due to the large number of parameters defining it.`
Is the main source of uncertainty due the number of parameters or the assumptions and the abstractions in modeling?

- [x] Third sentence needs a comma or a slight rework.

#### Statement of need
- [x] Exploration and search phases: maybe `search phase` better described as `exploitation` or `optimization` or `dimensioning`?
- [x] `A full PDOPT test case` I think this is not for just test cases.
- [x] `Thanks to the Exploration phase, the computational cost for design space exploration can be reduced by up to 80%`
Can this be a bit reworked or detail added. "thanks to exploration, exploration is faster..." does not sound right. Maybe "advanced exploration" or something.
- [x] The speedup cited in this work is the filtering method proposed in Spinelli, Anderson, et al (2022)?

#### State of the field
- [x] Review checklist requires a sort of "state of the field".  Consider adding this explicitly as a new section or implicitly in existing sections to address the point below.
- [x] Add how a comment how this problem is typically solved and possibly what software packages. Compare this software to typical methods.  Advantages and disadvantages

#### References
-  [x] Consider extending sources beyond self-citations
