+++ example/example_2/energy_management_experiment.py
@@ -54,7 +54,7 @@ import numpy as np
 from tqdm import tqdm
 
 from pdopt.data import DesignSpace, ExtendableModel
-from pdopt.exploration import  ProbabilisticExploration
+from pdopt.exploration import ProbabilisticExploration
 from pdopt.optimisation import Optimisation
 from pdopt.tools import generate_run_report
 
@@ -62,262 +62,247 @@ from HE_Model import model, postpro_run
 
 
 class Experiment(ExtendableModel):
-    
     def __init__(self, input_parameters, architecture, mission_file):
-        self.inp        = list(pd.read_csv(input_parameters)['name'])
-        self.arch       = architecture
-        self.arch0      = architecture
-        self.mission    = mission_file
-        
+        self.inp = list(pd.read_csv(input_parameters)["name"])
+        self.arch = architecture
+        self.arch0 = architecture
+        self.mission = mission_file
+
     def run(self, *args, **kwargs):
         # The input of the model is variable
         # we need to construct the energy management dataframe
-        
+
         X, parms = [], []
         en_mgm = []
-        
+
         # The TO/LND conditions will be single point only
         # logic to convert TO/LND into takeoff, climbout, final, landing
         for i in range(len(self.inp)):
+            if "TO" in self.inp[i]:
+                parms.append(self.inp[i].replace("TO", "takeoff"))
+                parms.append(self.inp[i].replace("TO", "climbout"))
+                X.append(args[i])
+                X.append(args[i])
 
-            if 'TO' in self.inp[i]:
-                parms.append(self.inp[i].replace('TO','takeoff'))
-                parms.append(self.inp[i].replace('TO','climbout'))
+            elif "LND" in self.inp[i]:
+                parms.append(self.inp[i].replace("LND", "final"))
+                X.append(args[i])
+
+            elif "LNDToAlternate" in self.inp[i]:
+                parms.append(self.inp[i].replace("LND", "final"))
+                parms.append(self.inp[i].replace("LNDToAlternate", "landing"))
                 X.append(args[i])
                 X.append(args[i])
-                
-            elif 'LND' in self.inp[i]:
-                parms.append(self.inp[i].replace('LND','final'))
-                X.append(args[i])
-                
-            elif 'LNDToAlternate' in self.inp[i]:
-                parms.append(self.inp[i].replace('LND','final'))
-                parms.append(self.inp[i].replace('LNDToAlternate','landing'))
-                X.append(args[i])
-                X.append(args[i])
-            
+
             else:
                 parms.append(self.inp[i])
                 X.append(args[i])
-                
-        
+
         # Stuff to introduce constraints over the x_positions
         all_x_vals = []
         x_vals = []
         old_seg = None
-        
+
         for i in range(len(parms)):
             # separate segment from type
-            segment, value = parms[i].split('_')
-            
+            segment, value = parms[i].split("_")
+
             if old_seg != segment and len(x_vals) > 0:
                 all_x_vals.append(x_vals)
 
             x_vals = [] if old_seg != segment else x_vals
 
-            if value[0] == 'h':
+            if value[0] == "h":
                 if len(en_mgm) == 0 or en_mgm[-1][0] != segment:
                     en_mgm.append([segment, 0, X[i]])
                 else:
                     en_mgm.append([segment, 1, X[i]])
-            elif value[0] == 'x':
+            elif value[0] == "x":
                 x_vals.append(X[i])
                 en_mgm[-1][1] = X[i]
-                
+
             old_seg = segment
         all_x_vals.append(x_vals)
 
-        
[…]
