+++ tests/test_optimisation.py
@@ -13,12 +13,16 @@ import pandas as pd
 
 @pytest.fixture
 def my_DesignSpace():
-    return data.DesignSpace.from_csv('test_input.csv', 'test_response.csv')
+    return data.DesignSpace.from_csv("test_input.csv", "test_response.csv")
+
 
 def dummy_function(x, y):
-    return {'obj'  : (x-5)**2 - y*5,
-            'con1' : (x-5)**2 - y*5,
-            'con2' :  11 - 0.5*x**2}
+    return {
+        "obj": (x - 5) ** 2 - y * 5,
+        "con1": (x - 5) ** 2 - y * 5,
+        "con2": 11 - 0.5 * x**2,
+    }
+
 
 @pytest.fixture
 def my_Model():
@@ -28,393 +32,511 @@ def my_Model():
 ## Test NNSurrogate
 @pytest.fixture
 def my_NNSurrogate(my_Model, my_DesignSpace):
-    return optimisation.NNSurrogate(my_Model, my_DesignSpace, 
-                                              set_id=0, debug=True)
+    return optimisation.NNSurrogate(my_Model, my_DesignSpace, set_id=0, debug=True)
+
 
 ## The debug flag fixes all the random states
 def test_NNSUrrogate_sample(my_NNSurrogate):
     # Check that it correctly sampled the same things
     my_NNSurrogate.sample()
-    
-    expected_X_test = np.array([[0.66897452, 1.        ],
-       [0.1876068 , 1.        ],
-       [0.91404481, 1.        ],
-       [0.50982793, 0.        ],
-       [0.08327291, 1.        ],
-       [0.26699526, 0.        ],
-       [0.7688592 , 0.        ],
-       [0.13703354, 0.        ],
-       [0.48450821, 0.        ],
-       [0.63535057, 0.        ],
-       [0.09508085, 0.        ],
-       [0.94989628, 1.        ],
-       [0.15825541, 1.        ],
-       [0.35966759, 1.        ],
-       [0.01861538, 0.        ],
-       [0.79755058, 0.        ],
-       [0.3003415 , 1.        ],
-       [0.55104381, 1.        ],
-       [0.0513041 , 1.        ],
-       [0.7183425 , 0.        ],
-       [0.93501122, 1.        ],
-       [0.7351483 , 0.        ],
-       [0.37991396, 1.        ],
-       [0.45792965, 1.        ],
-       [0.57615978, 0.        ],
-       [0.33627077, 0.        ],
-       [0.60473845, 1.        ],
-       [0.86263638, 1.        ],
-       [0.42373889, 0.        ],
-       [1.00832237, 0.        ],
-       [0.8318984 , 0.        ],
-       [0.22887941, 1.        ]])
-    
-    expected_X_train = np.array([[0.43177599, 1.        ],
-           [0.97354942, 0.        ],
-           [0.56810277, 1.        ],
-           [0.02741659, 0.        ],
-           [0.23473156, 1.        ],
-           [0.64798802, 0.        ],
-           [0.76601971, 1.        ],
-           [0.35210548, 0.        ],
-           [0.25348936, 1.        ],
-           [0.86058461, 0.        ],
-           [0.75536051, 1.        ],
-           [0.14736584, 0.        ],
-           [0.06747669, 1.        ],
-           [0.54804943, 0.        ],
-           [0.94025123, 1.        ],
-           [0.46102296, 0.        ],
-           [0.48315498, 1.        ],
-           [0.89817551, 0.        ],
-           [0.52582001, 1.        ],
-           [0.10964979, 0.        ],
-           [0.15765488, 1.        ],
-           [0.70119249, 0.        ],
-           [0.85044749, 1.        ],
-           [0.30750544, 0.        ],
[…]

[…]
