### 2024-01-24 spinjet/pdopt-code#3 @spinjet
Dear @jbussemaker @e-dub

I ran the code on a fresh installation and had the same FutureWarning issue.
It stems from when the file `mission.csv` is loaded in the code at line 336 of HE_model.py: https://github.com/spinjet/pdopt-code/blob/d2bf7d21ed99f6790ab0c4ec7ca927a0e4a6cc6b/example/example_1/HE_Model.py#L335

In the .csv file the column pDOH only contains zeros, hence Pandas interpreted it as a int64 type. As the degree of hybridisation is then defined in the function `generate_doh`, this is replaced with a float value from 0 to 1. It seems in previous versions of Pandas this recast didn't throw any Warnings, hence I never noticed it.

This has now been fiixed by ensuring the dtype of the pDOH column in the dataframe is set to float when the energy management strategy is created (see in https://github.com/spinjet/pdopt-code/blob/31f65670a6fa449833f975c7cba305950c74c2ad/example/example_1/HE_Model.py#L330):

```
   mission.loc[mission['tag'] == s, 'pDOH'] = None
   mission.loc[mission['tag'] == s, 'pDOH'] = tmp.astype(float)
```

I also took the opportunity to do a bit of linting in the example files.
