### 2024-01-24 spinjet/pdopt-code#2 @spinjet
Dear @jbussemaker

Helper functions have now been fully implemented as per commit: https://github.com/spinjet/pdopt-code/commit/31f65670a6fa449833f975c7cba305950c74c2ad

Specifically:
- Standard `DesignSpace` interface is Python lists with objects for parameters, objectives and constraints
- Helper function to load and create `DesignSpace` objects `from_csv` and `from_pickle` added
- Helper function to save `DesignSpace` objects `save_to_pickle` added
- Helper function to save results to .csv (`save_exploration_results` and `save_optimisation_results`)
- Helper functions to save and  load `ProbabilisticExploration` objects from pickle have been added (`from_pickle` and `save_to_pickle`).

Hope this is enough for the improvements.
