### 2024-01-24 spinjet/pdopt-code#1 @jbussemaker
@spinjet great progress here

One comment about the `.gitignore` file: please also remove all content from the repository that would be "ignored". For example, `build` and `dist` should not be present.
