### 2024-01-24 spinjet/pdopt-code#4 @spinjet
Dear @e-dub
Thank you for your comments and suggestions.
I will address these as also @jbussemaker had similar feedback for the paper.

> #### Summary
> * [ ]  `Contemporary...`
>   I do not know if only contemporary systems are characterized in this way. What is different now is that we are now trying better control/master/handle/design with this complexity, no?

I agree with your interpretation, perhaps in the search for brevity this meaning is not well communicated. My intended idea was that in the search of better optimality, more and more detail is present in the earlier design phases, hence we have to deal with complex systems more generally, be it an airframe or a consumer automobile. It's true historically complex systems have existed, but were limited to special applications (say Nuclear powerplants, Space systems and so on). So dealing with complexities early on, especially in the aeronautical industry, is a phenomenon of the last 20-30 years. For instance, aircraft design until the 1960s could be initiated by a single chief designer up to the preliminary stage. Nowadays this is almost impossible. I hope this conveys the story I had in mind, I can extend the work possibly.

> * [ ]  You are framing the context within the realm of uncertainty in the summary.  Do I understand correctly that the surrogate model is able to capture the design space and its uncertainty?  Or does the surrogate model work in a probabilistic way?  Can you please add a sentence to make this clearer?

The surrogate model captures uncertainty in the sense that it's a Gaussian Process and its being leveraged for calculating the probability a design point falls within the decision boundary of a requirement (i.e. the domain where the requirement is satisfied in the design space). This probability should be interpreted in a Bayesian way: it measures the level of confidence the requirement is satisfied. The advantage of this approach is twofold:

- Sets that fall at the edge of a decision boundary can be handled with less ambiguity as using Bayesian probabilities introduces a level of fuzziness on the boundary. During the exploration phase this is fine, as we are mainly concerned with eliminating what we are mostly sure it's unfeasible.
- It enables to leverage probabilistic logic for performing the set elimination process without IF-ELSE logic, that have to rely on extra assumptions on top of the design model. This is the major hypothesis of PDOPT where set-based elimination is carried out by extracting the knowledge embedded in the design model, rather than introducing new rules.

> * [ ]  `The design of such systems entails high uncertainty due to the large number of parameters defining it.`
>   Is the main source of uncertainty due the number of parameters or the assumptions and the abstractions in modelling?

Within PDOPT the source of uncertainty is epistemic: that is the designer does not know precisely which combination of design parameters satisfy the requirements set forward by the problem. Indeed, as explained above, the code leverages the probabilistic exploration to restrict the possible range of design sets to look for by eliminating those where there's an high confidence the requirements are not satisfied. So I reckon the sentence doesn't properly convey the intended concept, I'll correct it. What I had in mind was the number of parameters leave a lot of degrees of freedoms in the design problem hence the epistemic uncertainty from finding the right set of parameters.

> #### Statement of need
> * [ ]  Exploration and search phases: maybe `search phase` better described as `exploitation` or `optimization` or `dimensioning`?

In my PhD thesis I preferred the use of search phase because "optimisation" may be confused with the actual optimisation being carried out within each surviving design set. Since this step searches each set with a local optimisation problem to recover the actual design points, I used the word "search". Exploitation may be a suitable candidate as it's more familiar in the computational engineering and mathematical optimisation fields. Admittedly, in the codebase the word "optimisation" is used for the module carrying out the search phase, so perhaps I should reflect on which nomenclature to use to uniform it.

> * [ ]  `A full PDOPT test case` I think this is not for just test cases.

Apologies, what I meant a full run or PDOPT analysis. I'll correct it immediately.

> * [ ]  `Thanks to the Exploration phase, the computational cost for design space exploration can be reduced by up to 80%`
>   Can this be a bit reworked or detail added. "thanks to exploration, exploration is faster..." does not sound right. Maybe "advanced exploration" or something.
> * [ ]  The speedup cited in this work is the filtering method proposed in Spinelli, Anderson, et al (2022)?

Yes, the computational cost saving is for the overall design space exploration. In other words we avoid having to run an optimisation in areas of the design space that would not yield feasible design. This was originally shown as a proof of concept in Spinelli, Anderson et al (2022) as you correctly point out. I see that the sentence structure of `Thanks to the Exploration [...] by up to 80%` is confusing and may lead the reader to confuse the overall design space exploration task with the "exploration step" of PDOPT. This is going to be reworked.

> #### State of the field
> * [ ]  Review checklist requires a sort of "state of the field".  Consider adding this explicitly as a new section or implicitly in existing sections to address the point below.
> * [ ]  Add how a comment how this problem is typically solved and possibly what software packages. Compare this software to typical methods.  Advantages and disadvantages

@jbussemaker also suggested to include this section. I originally thought providing references to other publications may be enough in case a reader is interested in deepen its understanding of this topic, but I agree it's necessary. Perhaps this was missing from the JOSS requirements when I made the submission? I plan to include this section with also a small reference to theory.

> #### References
> * [ ]  Consider extending sources beyond self-citations

I agree, this was also pointed out by @jbussemaker. I plan to provide some references to the theory that underlies the concepts PDOPT leverages and similar approaches for design space exploration.

From your questions and suggestions I recognise the paper doesn't provide enough context on the theory and ideas behind PDOPT. I am really thankful for your questions because they elucidated some gaps in the paper which I completely overlooked during the submission. I will work on these and include some figures which help visualise what the framework is doing.
