### 2024-01-24 0b48bf2
Add mkdocs documentation
Added a stub for readthedocs supported by mkdocs
