+++ docs/requirements.txt
@@ -1,0 +1,2 @@
+# Defining the exact version will make sure things don't break
+mkdocs==1.5.3
