+++ .github/workflows/pytest.yml
@@ -1,14 +1,36 @@
-name: pytest
+name: Lint and Test
 on: [push]
     
 jobs:
+
+  run-linter:
+    name: Run linter
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check out Git repository
+        uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.9
+
+      - name: Install Python dependencies
+        run: pip install black
+
+      - name: Run linters
+        uses: wearerequired/lint-action@v2
+        with:
+          black: true
+
   test:
     strategy:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest]
         python-version: ["3.9", "3.10", "3.11", "3.12"]
-    name: test
+    name: Test
     runs-on: ${{ matrix.os }}
     steps:
      - name: Checkout code
