### 2024-01-26 41bca89
Updated paper.md
Added WIP State of the Field section
