+++ paper/paper.bib
@@ -79,3 +79,148 @@ chapter = {},
 pages = {},
 doi = {10.2514/6.2023-0837},
 }
+
+@article{MartinsLambe2013,
+author = {Martins, Joaquim R. R. A. and Lambe, Andrew B.},
+title = {Multidisciplinary Design Optimization: A Survey of Architectures},
+journal = {AIAA Journal},
+volume = {51},
+number = {9},
+pages = {2049-2075},
+year = {2013},
+doi = {10.2514/1.J051895},
+}
+
+@article{saves2024smt,
+        author = {P. Saves and R. Lafage and N. Bartoli and Y. Diouane and J. Bussemaker and T. Lefebvre and J. T. Hwang and J. Morlier and J. R. R. A. Martins},
+        title = {{SMT 2.0: A} Surrogate Modeling Toolbox with a focus on Hierarchical and Mixed Variables Gaussian Processes},
+        journal = {Advances in Engineering Sofware},
+        year = {2024},
+        volume = {188},
+        pages = {103571},
+        doi = {https://doi.org/10.1016/j.advengsoft.2023.103571}
+		}
+
+@article{pymoo,
+    author={J. {Blank} and K. {Deb}},
+    journal={IEEE Access},
+    title={pymoo: Multi-Objective Optimization in Python},
+    year={2020},
+    volume={8},
+    number={},
+    pages={89497-89509},
+}
+
+@article{Gray2019a,
+    Author = {Justin S. Gray and John T. Hwang 
+              and Joaquim R. R. A. Martins 
+              and Kenneth T. Moore and Bret A. Naylor},
+    Doi = {10.1007/s00158-019-02211-z},
+    Journal = {Structural and Multidisciplinary Optimization},
+    Month = {April},
+    Number = {4},
+    Pages = {1075--1104},
+    Title = {{OpenMDAO}: An open-source framework 
+             for multidisciplinary design, analysis, 
+             and optimization},
+    Volume = {59},
+    Year = {2019}}
+
+@article{DebNSGA:2002,
+  author={Deb, K. and Pratap, A. and Agarwal, S. and Meyarivan, T.},
+  journal={IEEE Transactions on Evolutionary Computation}, 
+  title={A fast and elitist multiobjective genetic algorithm: NSGA-II}, 
+  year={2002},
+  volume={6},
+  number={2},
+  pages={182-197},
+  keywords={Genetic algorithms;Sorting;Computational complexity;Evolutionary computation;Computational modeling;Testing;Decision making;Associate members;Diversity reception;Constraint optimization},
+  doi={10.1109/4235.996017}}
+  
+@inbook{Deb2008,
+author={Kalyanmoy Deb},
+editor={J Branke and K Deb and and K Miettinen and R S{\l}owi{\'{n}}ski},
+title={Introduction to Evolutionary Multiobjective Optimization},
+booktitle={Multiobjective Optimization: Interactive and Evolutionary Approaches},
+year={2008},
+publisher={Springer Berlin Heidelberg},
+address={Berlin, Heidelberg},
+pages={59--96},
+isbn={978-3-540-88908-3},
+doi={10.1007/978-3-540-88908-3_3},
+}
+
+@article{Forrester2009,
+title = {Recent advances in surrogate-based optimization},
+journal = {Progress in Aerospace Sciences},
+volume = {45},
+number = {1},
+pages = {50-79},
+year = {2009},
+issn = {0376-0421},
+doi = {https://doi.org/10.1016/j.paerosci.2008.11.001},
+author = {Alexander I.J. Forrester and Andy J. Keane},
+}
+
+@inproceedings{Kennedy1995,
[…]
