+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 # Defining the exact version will make sure things don't break
 mkdocs==1.5.3
+mkdocstrings[python]>=0.18
