+++ mkdocs.yml
@@ -3,6 +3,9 @@ site_name: PDOPT Documentation
 nav:
   - Home: index.md
   - About: about.md
+  - pdopt.data : data.md
 
 theme: readthedocs
 
+plugins:
+  - mkdocstrings 
