### 2024-01-26 #6110 @spinjet
Dear @kyleniemeyer

I've begun addressing the points @jbussemaker and @e-dub have made on the paper.
Currently I've added a State of the field section to provide some context on similar approaches for design space exploration.

Paper corrections is still WIP, and unfortunately next week I wont' be able to make any progress until Friday due to a workshop that will take most of the week.

Please let me know if any suggestions come through.
