### 2024-01-26 spinjet/pdopt-code#4 @spinjet
Added a almost completed section for State of the Field with relevant references.
