### 2024-01-26 7ea78b9
Updated paper
Added WIP State of the Field section
