+++ paper/paper.bib
@@ -224,3 +224,15 @@ issn = {0376-0421},
 doi = {https://doi.org/10.1016/j.paerosci.2017.11.003},
 author = {Raul Yondo and Esther Andrés and Eusebio Valero},
 }
+
+@article{singer2009set,
+	title={What is set-based design?},
+	author={Singer, David J and Doerry, Norbert and Buckley, Michael E},
+	journal={Naval Engineers Journal},
+	volume={121},
+	number={4},
+	pages={31--43},
+	year={2009},
+	publisher={American Society of Naval Engineers},
+	doi={10.1111/j.1559-3584.2009.00226.x}
+}
