### 2024-01-30 4dc9ea8
Updated paper
Completed the State of the Field section, added figures.
