+++ paper/paper.bib
@@ -236,3 +236,58 @@ author = {Raul Yondo and Esther Andrés and Eusebio Valero},
 	publisher={American Society of Naval Engineers},
 	doi={10.1111/j.1559-3584.2009.00226.x}
 }
+
+@article{ward1993sbd,
+	author = {Allen C. Ward and Warren P. Seering},
+	title = {Quantitative inference in a mechanical design 'compiler'},
+	journal = {Journal of Mechanical Design},
+	volume = {115},
+	number = {1},
+	pages = {29-35},
+	year = {1993},
+	month = {03},
+	doi = {10.1115/1.2919320},
+}
+
+
+@inproceedings{sobek1996sbdToyota,
+	title={Principles from Toyota’s set-based concurrent engineering process},
+	author={Sobek, Durward K and Ward, Allen C},
+	booktitle={International Design Engineering Technical Conferences and Computers and Information in Engineering Conference},
+	volume={97607},
+	pages={V004T04A035},
+	year={1996},
+	organization={American Society of Mechanical Engineers},
+	doi = {10.1115/96-DETC/DTM-1510},
+}
+
+@article{mckenney2011adapting,
+	title={Adapting to changes in design requirements using set-based design},
+	author={McKenney, Thomas A and Kemink, Lauren F and Singer, David J},
+	journal={Naval Engineers Journal},
+	volume={123},
+	number={3},
+	pages={67--77},
+	year={2011},
+	publisher={American Society of Naval Engineers},
+	doi={10.1111/j.1559-3584.2011.00331.x}
+}
+
+@article{ADOPT_paper,
+	title={ADOPT: An augmented set-based design framework with optimisation},
+	author={Georgiades, Alex and Sharma, Sanjiv and Kipouros, Timoleon and Savill, Mark},
+	journal={Design Science},
+	volume={5},
+	pages={e4},
+	year={2019},
+	publisher={Cambridge University Press},
+	doi = {10.1017/dsj.2019.1}
+}
+
+@book{bernardo2009bayesian,
+  title={Bayesian theory},
+  author={Bernardo, Jos{\'e} M and Smith, Adrian FM},
+  volume={405},
+  year={2009},
+  publisher={John Wiley \& Sons}
+}
