+++ .github/workflows/pytest.yml
@@ -1,29 +1,7 @@
-name: Lint and Test
+name: Tests
 on: [push]
     
 jobs:
-
-  run-linter:
-    name: Run linter
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Check out Git repository
-        uses: actions/checkout@v3
-
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: 3.9
-
-      - name: Install Python dependencies
-        run: pip install black
-
-      - name: Run linters
-        uses: wearerequired/lint-action@v2
-        with:
-          black: true
-
   test:
     strategy:
       fail-fast: false
