+++ .github/workflows/draft-pdf.yml
@@ -1,3 +1,4 @@
+name: Compile Paper
 on: [push]
 
 jobs:
