+++ paper/paper.md
@@ -42,7 +42,7 @@ Examples of libraries useful for performing the design of experiments are the Qu
 
 The framework presented in this paper attempts to combine the two approaches in a single framework through the application of the principles of Set-Based Design (SBD) and Bayesian probability. SBD is a design practice that focuses on narrowing down the input design space by eliminating the candidate designs that do not satisfy the needs and requirements [@singer2009set]. It was originally developed by Ward [@ward1993sbd] and it is commonly associated with the Toyota Production System [@sobek1996sbdToyota]. Figure \autoref{fig:sbd_process} compares SBD with the traditional iterative design process. Focusing on eliminating the unfeasible and undesirable designs from the candidate pool rather than selecting one to refine leads to a more robust design cycle [@mckenney2011adapting].
 
-![Comparison between Set-Based Design and traditional design.\label{fig:sbd_process}](sbd_process.png)
+![Comparison between Set-Based Design and traditional design.\label{fig:sbd_process}](sbd_process.png){ width=75% }
 
 SBD is used in `PDOPT` for performing an initial assesment of the design space and restriction to the most promising portions, which are then evaluated with local MDO problems. Georgiades previously developed a framework combining SBD and MDO named `ADOPT` [@ADOPT_paper], of which `PDOPT` is a development. The difference is in the set-elimination process. `ADOPT` used expert-defined rules that mapped the input parameters to the quantities of interest. This approach is robust in case of well understood design problems, but limited for unconventional systems, of which there is no best practices to draw from. `PDOPT` overcomes this limitation by applying Bayesian probability [@bernardo2009bayesian] as a selection criterion and assuming the underlying MDO model is a source of knowledge for the set elimination process. By casting the requirements in a probabilistic statement (i.e. "What is the probability it is satisfied?") and sampling in each set using surrogate models, it is possible to estimate the likelihood a set can satisfy all the requirements simultaneously and, therefore, worthy of further analysis. The advantage of this methodology is not having to rely on additional hardcoded rules on top of the implicit assumptions in the MDO model. 
 
@@ -65,8 +65,9 @@ The aggregation of these design points yields the global Pareto front with feasi
 
 The framework presented in this is built upon Set-Based Design and ... (WIP)
 
-![Caption for example figure.\label{fig:example}](sbd_process.png)
-![Caption for example figure.\label{fig:example}](set_selection.png)
+![Decision boundary and domain of a generic requirement.\label{fig:db_1}](decision_boundary.png){ width=33% }
+![Probabilistic decision boundary.\label{fig:db_2}](db_probabilistic.png){ width=33% }
+![Probabilistic decision boundary and set boundaries.\label{fig:db_3}](db_sets.png){ width=33% }
 
 # Availability
 
