+++ paper/paper.md
@@ -25,7 +25,9 @@ bibliography: paper.bib
 
 # Summary
 
-Contemporary engineering systems are characterised by many components and complex interactions between them. The design of such systems entails high uncertainty due to the large number of parameters defining it. An approach to manage it is exploring and evaluating as many alternatives as possible, before committing to a specific solution. The Python package `PDOPT` aims to provide this capability without the high computational cost associated with the factorial-based design of experiments methods. By exploiting a probabilistic machine learning model, the code identifies the areas of the design space which are most promising for the requirements provided by the user. The result is a large number of feasible design points, aiding the designer in understanding the behaviour of the system under design and selecting the desired configuration for further development.
+Contemporary engineering design is characterised by products and systems with increasing complexity coupled with tighter requirements and tolerances. This leads to high epistemic uncertainty due to numerous possible configurations and a high number of design parameters. Set-Based Design is a methodology capable of handling these design problems, by exploring and evaluating as many alternatives as possible, before committing to a specific solution. 
+
+The Python package `PDOPT` aims to provide this capability without the high computational cost associated with the factorial-based design of experiments methods. Additionally, `PDOPT` performs the requirement mapping without explicit rule definition. Instead, it utilizes a probabilistic machine learning model to identify the areas of the design space most promising for user-provided requirements. This yields a plethora of feasible design points, assisting designers in understanding the system behaviour and selecting the desired configurations for further development.
 
 # State of the field
 
@@ -51,22 +53,35 @@ SBD is used in `PDOPT` for performing an initial assesment of the design space a
 
 `PDOPT`, short for Probabilistic Design and OPTimisation, is a Python package for design space exploration of systems under design. It implements a set-based approach for mapping the requirements to the design space, using a probabilistic surrogate model trained on the provided design model. This procedure ensures to identification of the best candidate areas of the design space with the minimum number of assumptions of the design of the system.
 
-The API of `PDOPT` was designed as a library with class-based interfaces between the components of the framework. This ensures both flexibility and transparency, as the user can inspect the main data structure between the phases of the framework. A full `PDOPT` analysis consists of two phases: the Exploration phase and the Search phase. These are shown in Figure \autoref{fig:pdopt_chart}. The first one surveys the design space to identify the areas that are most likely to satisfy the constraints over the quantities of interest of the model. This is carried out by breaking down the design space into a hypercube of parameters’ levels (named sets), and rapidly evaluating them with the probabilistic surrogate model. The second phase introduces a multi-objective optimisation problem in each surviving design space area for recovering the individual design points. The result is multiple local Pareto fronts, one for each set.
+The API of `PDOPT` was designed as a library with class-based interfaces between the components of the framework. This ensures both flexibility and transparency, as the user can inspect the main data structure between the phases of the framework. A full `PDOPT` analysis consists of two phases: the Exploration phase and the Search phase. These are shown in Figure \autoref{fig:pdopt_chart}. The first one surveys the design space to identify the areas that are most likely to satisfy the constraints over the quantities of interest of the model. The second phase introduces a multi-objective optimisation problem in each surviving set for recovering the individual design points. The result is multiple local Pareto fronts, one for each set.
 
 ![Overall architecture of PDOPT.\label{fig:pdopt_chart}](pdopt_process.png)
 
[…]
