### 2024-02-05 spinjet/pdopt-code#4 @spinjet
Dear @e-dub.

I've made the improvements on the paper that you have addressed.
I've added a State of the Field section, introduced more references, and made improvements in the section you discussed.
