### 2024-02-05 spinjet/pdopt-code#1 @spinjet
Dear @jbussemaker.
I've completed the paper improvements.
Added a State of the Field section providing more context on the code and comparison with other frameworks/approaches. Introduced more references, including to the methods used in PDOPT. These also address the points @e-dub made in https://github.com/spinjet/pdopt-code/issues/4.

Hope this is satisfactory from the paper side, I'll focus on publishing in PyPI and having an online version of the documentation (especially for the API).
