### 2024-02-05 #6110 @spinjet
Dear @kyleniemeyer
I've completed the paper improvements, as required by the reviewers.
https://github.com/spinjet/pdopt-code/issues/1#issuecomment-1927242052

I think all the important steps for publication are completed.
I'm going to have the code published on pyPI and have an online documentation for the code to polish the submission.
Let me know if anything else is required.
