### 2024-02-05 #6110 @jbussemaker
> :point_right::page_facing_up: [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06110/joss.06110/10.21105.joss.06110.pdf) :page_facing_up: [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.06110/joss.06110/10.21105.joss.06110.pdf) :page_facing_up: :point_left:

@spinjet this paper version doesn't have a state of the field section, and still only a few self-references. Were your improvements committed and pushed?
