### 2024-02-06 #6110 @jbussemaker
@spinjet thank you for the updates to the paper, I think it has really improved the quality and made it more relevant for readers.

The only thing is that I wouldn't call the Surrogate Modeling Toolbox (SMT) a development of UoM: it's quite a large consortium, and the latest developments have been more lead by ONERA. Also, its purpose is not only DOE's, but also surrogate modeling (focusing on Gaussian Processes, which are relevant to PDOPT) of course ;)

After this, I'll be happy to recommend publication.
