+++ src/pdopt/data.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Created on Fri Oct 15 11:31:14 2021
-
-Module that contains all the data structures used within PDOPT.
-
+Module that contains all the data structures used within `PDOPT`.
 """
 
 # Standard Library Imports
@@ -74,7 +71,7 @@ class Response:
 
 
 class ExtendableModel:
-    # Model Object that can be extended and used by the library
+    '''Model Object that can be extended and used by the library'''
     def __init__(self):
         pass
 
@@ -266,9 +263,36 @@ class Constraint(Response):
 
 
 class DesignSet:
+    '''
+    A class to represent a Design Set.
+    
+    Attributes:
+        id (int): Unique id of the set.
+        parameter_levels_dict (dict[P]): .
+        parameters (list[Parameter]): List of parameters.
+        objectives (list[Objective]): List of objectives.
+        constraints (list[Constraint]): List of constraints.
+        n_par (int): Number of paramters.
+        par_names (list[str]): List of the parameter names.
+        obj_names (list[str]): List of objective names.
+        con_names (list[str]): List of constraint names.
+        sets (list[DesignSet]): List of sets within the design space.
+    '''    
     _ids = count(0)
 
     def __init__(self, input_parameter_levels, response_parameters):
+        """
+        
+
+        Args:
+            input_parameter_levels (TYPE): DESCRIPTION.
+            response_parameters (TYPE): DESCRIPTION.
+
+        Returns:
+            None.
+
+        """
+        
         self.id = next(self._ids)
 
         self.parameter_levels_dict = input_parameter_levels
@@ -316,18 +340,57 @@ class DesignSet:
             return s1
 
     def get_discarded_status(self):
+        """
+        Get the discarded status of the set.
+
+        Returns:
+            bool: Discarded status.
+
+        """
+        
         return self.is_discarded
 
     def get_P(self):
+        """
+        Get the overall probability of the set.
+
+        Returns:
+            float: Probability value.
+
+        """
+        
         return self.P
 
     def get_response_P(self, response_id=None):
+        """
+        Returns the probability of a response or all of them.
+
+        Args:
+            response_id (int, optional): Get the probability of a response. If set to None, get all the responses.
+
+        Returns:
+            float: Probability value.
+
+        """
+        
         if response_id:
             return self.P_responses[response_id]
         else:
             return self.P_responses
 
     def set_responses_P(self, response_name, P_response):
+        """
+        Updates adds the probability of the response and updated the global probability.
+
+        Args:
+            response_name (str): Name of the response (Must be a constraint or objective).
+            P_response (float): Probability value of the response.
+
+        Returns:
+            None.
+
+        """
+        
         # Add the new response result, and updates the total probability
         self.P_responses.update({response_name: P_response})
 
@@ -336,12 +399,45 @@ class DesignSet:
             self.P *= self.P_responses[k]
 
     def set_as_discarded(self):
+        """
+        Set the set as discarded.
+
+        Returns:
+            None.
+
+        """
+        
         self.is_discarded = True
 
     def set_optimisation_problem(self, opt_problem):
+        """
+        Set the optimisation problem from the Optimisation library
+
+        Args:
+            opt_problem (pymoo.Problem): Optimisation problem.
+
+        Returns:
+            None.
+
+        """
+        
         self.optimisation_problem = opt_problem
 
     def sample(self, n_samples, parameters_list, debug=False):
+        """
+        Sample designs within the set using Latin Hypercube
+
+        Args:
+            set_id (int): ID of the set.
+            n_samples (int): Number of samples to be generated.
[…]
