+++ docs/index.md
@@ -1,20 +1,21 @@
-# Welcome to MkDocs
+# PDOPT: Probabilistic Design space exploration and OPTimisation
+
+A `python` framework for set-based design space exploration without explicit elimination rules. It impements a set-based approach for mapping the requirements on the design space, using a probabilistic surrogate model trained on the provided design model. This procedure ensures to identify the best candidate areas of the design space with the minimum number of assumptions of the design of the system.
+
 
 For full documentation visit [mkdocs.org](https://www.mkdocs.org).
 
-## Commands
+## Installation
 
-* `mkdocs new [dir-name]` - Create a new project.
-* `mkdocs serve` - Start the live-reloading docs server.
-* `mkdocs build` - Build the documentation site.
-* `mkdocs -h` - Print help message and exit.
 
-## Project layout
 
-    mkdocs.yml    # The configuration file.
-    docs/
-        index.md  # The documentation homepage.
-        ...       # Other markdown pages, images and other files.
+## Community Guidelines
 
-## Example Reference
+This software is released with the MIT license and comes as is, without warranty of any kind. It is maintained by the author, but anyone is free to contribute. Users can open a ticket on GitHub to provide feeback on bugs or contributions.
+
+## Acknowledgements
+
+This software was developed within [Project FutPrInt50](https://futprint50.eu/), with EU Horizon 2020 Grant No. 875551. The authors wants to thank all the researchers in the project who cotnributed with their input to shape the framework.
+
+## References
 
