+++ docs/data.md
@@ -1,19 +1,6 @@
 # Documentation for `pdopt.data`
 
 
-## ::: pdopt.data.DesignSpace
-	handler : python
-	options:
-		members:
-			- __init__
-			- from_csv
-			- from_pickle
-			- save_to_pickle
-			- get_exploration_results
-			- get_optimum_results
-			- set_discard_status
-			- save_exploration_results
-			- save_optimisation_results
-			- sample_from_set
-		show_source: false
-		
+## ::: pdopt.data
+
+
