### 2024-02-07 a012baa
Updated paper, adding online documentation
Added docstrings for the data.py module to automatically build online API
