+++ paper/paper.md
@@ -38,9 +38,9 @@ In the conceptual phase, optimisation can be employed to understand which combin
 
 Software libraries developed for optimisation in Python are the OpenMDAO framework, which focuses on multi-disciplinary optimisation [@Gray2019a], the Object-Oriented `pyOpt` [@Pyopt], the `DEAP` evolutionary optimisation library [@DEAP], and `pymoo` multi-objective optimisation library [@pymoo]. These libraries are open-source and general-purpose.
 
-While effective in finding the desired set of design points, optimisation methods are required to properly set up the problem to deliver meaningful results. This may be difficult at the beginning of the design procedure, as needs may be not fully defined, and the behaviour of the model not be fully understood. Design of Experiments methods can assist in initiating this process through sampling and evaluation of the design space to obtain an understanding of the system response. Traditionally this approach is carried out through factorial sampling, where all possible combinations of extreme parameter values are tested [@MontgomeryDE]. A response surface would then be built for analysing the response of the system inside the interval. This approach is feasible when the number of variables is low, as the number of experiments to be carried out grows exponentially. In the computational engineering field, quasi-random sampling is preferred, with Latin Hypercube sampling and Sobol sequences being most used [@Yondo2018]. These allow for training surrogate models for design exploration with a lower number of points in cases where factorial design would require an unfeasible amount.
+While effective in finding the desired set of design points, optimisation methods require to properly set up the problem to deliver meaningful results. This may be difficult at the beginning of the design procedure, as needs may be not fully defined, and the behaviour of the model not be fully understood. Design of Experiments (DoE) methods can assist in initiating this process through sampling and evaluation of the design space to obtain an understanding of the system response. Traditionally this approach is carried out through factorial sampling, where all possible combinations of extreme parameter values are tested [@MontgomeryDE]. A response surface would then be built for analysing the response of the system inside the interval. This approach is feasible when the number of variables is low, as the number of experiments to be carried out grows exponentially. In the computational engineering field, quasi-random sampling is preferred, with Latin Hypercube sampling and Sobol sequences being most used [@Yondo2018]. These allow for training surrogate models for design exploration with a lower number of points in cases where factorial design would require an unfeasible amount.
 
-Examples of libraries useful for performing the design of experiments are the Quasi-Monte Carlo sampling module `scipy.qmc` and the University of Michigan’s Surrogate Modelling Toolbox [@saves2024smt]. Unlike an optimisation problem, there is no fixed procedure to obtain the mapping of the requirements on the input parameters. The designer is free to play with the model, with the aid of visualisation methods to understand the problem at hand.
+Currently there is not a specific Python library to perform DoE-based design space exploration. Unlike an optimisation problem, there is no fixed procedure to obtain the mapping of the requirements on the input parameters. The designer is free to play with the model, with the aid of visualisation methods to understand the problem at hand. Examples of libraries useful for performing DoE sampling are the Quasi-Monte Carlo sampling module `scipy.qmc` and the Surrogate Modeling Toolbox (SMT) [@saves2024smt]. SMT also provides methods to build response surfaces from the sampled data, using both radial-basis function (RBF) interpolation and Kriging (Gaussian Process Regression). Alternatively, scipy provides a RBF interpolation function in `scipy.interpolate.Rbf`, while Kriging models can be built using `scikit-learn.gaussian_process` module [@pedregosa2011scikit] or with the library `GPpy` [@gpy2014].
 
 ![Comparison between Set-Based Design and traditional design.\label{fig:sbd_process}](sbd_process.png){ width=75% }
 
[…]
