+++ paper/paper.bib
@@ -291,3 +291,20 @@ author = {Raul Yondo and Esther Andrés and Eusebio Valero},
   year={2009},
   publisher={John Wiley \& Sons}
 }
+
+@article{pedregosa2011scikit,
+  title={Scikit-learn: Machine learning in Python},
+  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
+  journal={Journal of machine learning research},
+  volume={12},
+  number={Oct},
+  pages={2825--2830},
+  year={2011}
+}
+
+@Misc{gpy2014,
+  author =   {{GPy}},
+  title =    {{GPy}: A Gaussian process framework in python},
+  howpublished = {\url{http://github.com/SheffieldML/GPy}},
+  year = {since 2012}
+}
