+++ src/pdopt/data.py
@@ -79,6 +79,25 @@ class ExtendableModel:
     # whose keys are the reponse names defined in the responses.csv
     # file.
     def run(self, *args: list[float]) -> dict[str, float]:
+        """
+        The run() method has to be overloaded with the evaluation function required to run the
+        analysis. Input parameters must be a list in the order of the paramters passed to the
+        DesignSpace object. Output must be a dictionary containing for keyword the names of
+        the constraints and objectives as defined in the DesignSpace object.
+
+        Parameters
+        ----------
+        *args : list[float]
+            A list containing the input quantities, in the same order as in the DesignSpace object.
+
+        Returns
+        -------
+        dict[str, float]
+            A dicitonary containing the outputs (constraints and objectives) with keywords matching
+            those of the object.
+
+        """
+        
         pass
 
 
@@ -86,14 +105,96 @@ class ExtendableModel:
 
 
 class ContinousParameter(Parameter):
+    '''
+    A class to represent a Continous Parameter.
+    
+    Attributes:
+        id (int): 
+            Unique id of the parameter.
+        name (str): 
+            Name of the parameter.
+         n_levels (int): 
+             Number of levels of the continous parameter.
+        lb (float):
+            Lower bound value of the continous parameter.
+        ub (float):
+            Upper bound value of the continous parameter.
+        ranges (list[(float, float)]):
+            List of the bounds of each level.
+        uq_dist (str):
+            Type of uncertainty distribution to be applied to this parameter.
+            Options are "norm", "uniform" and "triang".
+        uq_var_l (float):
+            Lower percentile variation from the UQ distribution mean
+        uq_var_u (float):
+            Upper percentile variation from the UQ distribution mean
+    ''' 
+    
     def get_bounds(self):
+        """
+        Returns a tuple with the continous parameter bounds
+
+        Returns
+        -------
+        float
+            Lower bound of the continous parameter.
+        float
+            Upper bound of the continous parameter.
+
+        """
+        
         return self.lb, self.ub
 
     def get_level_bounds(self, level):
+        """
+        Returns a tuple containing the bounds of the selected level.
+
+
+        Parameters
+        ----------
+        level : int
+            N-th selected level.
+
+        Returns
+        -------
+        float
+            Lower bound of the selected level.
+        float
+            Upper bound of the selected level.
+
+        """
+        
         assert level < self.n_levels, "Selected level is above total number of levels"
         return self.ranges[level], self.ranges[level + 1]
 
     def __init__(self, name, lb, ub, n_levels, uq_dist, uq_var_l, uq_var_u):
+        """
+        Initialise the Continous Parameter object.
+
+        Parameters
+        ----------
+        name (str): 
+            Name of the parameter.
+        lb (float):
+            Lower bound value of the continous parameter.
+        ub (float):
+            Upper bound value of the continous parameter.
+        n_levels (int): 
+            Number of levels of the continous parameter.
+        uq_dist (str):
+            Type of uncertainty distribution to be applied to this parameter.
+            Options are "norm", "uniform" and "triang".
+        uq_var_l (float):
+            Lower percentile variation from the UQ distribution mean
+        uq_var_u (float):
+            Upper percentile variation from the UQ distribution mean
+
+        Returns
+        -------
+        None.
+
+        """
+        
         super().__init__(name)
 
         assert lb < ub, "Lower Bound value higher than Upper Bound value"
@@ -117,6 +218,24 @@ class ContinousParameter(Parameter):
         return s1 + s2 + s3 + ")\n"
 
     def sample(self, n_samples, level=None):
+        """
+        Sample within the entire continuous parameter or in a level.
+
+        Parameters
+        ----------
+        n_samples : int
+            Number of samples.
[…]
