+++ src/pdopt/data.py
@@ -85,16 +85,14 @@ class ExtendableModel:
         DesignSpace object. Output must be a dictionary containing for keyword the names of
         the constraints and objectives as defined in the DesignSpace object.
 
-        Parameters
-        ----------
-        *args : list[float]
-            A list containing the input quantities, in the same order as in the DesignSpace object.
+        Args:
+            *args : list[float]
+                A list containing the input quantities, in the same order as in the DesignSpace object.
 
-        Returns
-        -------
-        dict[str, float]
-            A dicitonary containing the outputs (constraints and objectives) with keywords matching
-            those of the object.
+        Returns:
+            dict[str, float]
+                A dicitonary containing the outputs (constraints and objectives) with keywords matching
+                those of the object.
 
         """
         
@@ -134,12 +132,11 @@ class ContinousParameter(Parameter):
         """
         Returns a tuple with the continous parameter bounds
 
-        Returns
-        -------
-        float
-            Lower bound of the continous parameter.
-        float
-            Upper bound of the continous parameter.
+        Returns:
+            float
+                Lower bound of the continous parameter.
+            float
+                Upper bound of the continous parameter.
 
         """
         
@@ -150,17 +147,15 @@ class ContinousParameter(Parameter):
         Returns a tuple containing the bounds of the selected level.
 
 
-        Parameters
-        ----------
-        level : int
-            N-th selected level.
+        Args:
+            level : int
+                N-th selected level.
 
-        Returns
-        -------
-        float
-            Lower bound of the selected level.
-        float
-            Upper bound of the selected level.
+        Returns:
+            float
+                Lower bound of the selected level.
+            float
+                Upper bound of the selected level.
 
         """
         
@@ -171,27 +166,25 @@ class ContinousParameter(Parameter):
         """
         Initialise the Continous Parameter object.
 
-        Parameters
-        ----------
-        name (str): 
-            Name of the parameter.
-        lb (float):
-            Lower bound value of the continous parameter.
-        ub (float):
-            Upper bound value of the continous parameter.
-        n_levels (int): 
-            Number of levels of the continous parameter.
-        uq_dist (str):
-            Type of uncertainty distribution to be applied to this parameter.
-            Options are "norm", "uniform" and "triang".
-        uq_var_l (float):
-            Lower percentile variation from the UQ distribution mean
-        uq_var_u (float):
-            Upper percentile variation from the UQ distribution mean
+        Args:
+            name (str): 
+                Name of the parameter.
+            lb (float):
+                Lower bound value of the continous parameter.
+            ub (float):
+                Upper bound value of the continous parameter.
+            n_levels (int): 
+                Number of levels of the continous parameter.
+            uq_dist (str):
+                Type of uncertainty distribution to be applied to this parameter.
+                Options are "norm", "uniform" and "triang".
+            uq_var_l (float):
+                Lower percentile variation from the UQ distribution mean
+            uq_var_u (float):
+                Upper percentile variation from the UQ distribution mean
 
-        Returns
-        -------
-        None.
+        Returns:
+            None.
 
         """
         
@@ -221,18 +214,16 @@ class ContinousParameter(Parameter):
         """
         Sample within the entire continuous parameter or in a level.
 
-        Parameters
-        ----------
-        n_samples : int
-            Number of samples.
-        level : int, optional
-            N-th level to sample in. If None, sample in the entire range.
-            The default is None.
+        Args:
+            n_samples : int
+                Number of samples.
+            level : int, optional
+                N-th level to sample in. If None, sample in the entire range.
+                The default is None.
 
-        Returns
-        -------
-        numpy.ndarray
-            Array of random samples of length `n_samples`.
[…]
