+++ src/pdopt/optimisation.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Created on Mon Oct 31 10:05:27 2022
+This module contains classes and functions used for the search phase.
 
 """
 
@@ -54,7 +54,7 @@ t_max = 60 * 60
 
 @contextlib.contextmanager
 def tqdm_joblib(tqdm_object):
-    """Context manager to patch joblib to report into tqdm progress bar given as argument"""
+    #Context manager to patch joblib to report into tqdm progress bar given as argument
 
     class TqdmBatchCompletionCallback(joblib.parallel.BatchCompletionCallBack):
         def __call__(self, *args, **kwargs):
@@ -76,7 +76,53 @@ def tqdm_joblib(tqdm_object):
 
 
 class NNSurrogate(Problem):
+    '''
+    Class that represents a deterministic optimisation problem, using a 
+    neural-network surrogate model for function evaluation. 
+    It is locally trained before use on the DesignSet it is part of. 
+    This class wraps the pymoo.core.problem.Problem class.
+    
+    Attributes:
+        design_space (pdopt.data.DesignSpace): 
+            The design space object of the problem.
+        var (list[pdopt.data.Parameter]): 
+            List containing the input parameter objects.
+        obj (list[pdopt.data.Objective]): 
+            List containing the objective objects.
+        cst (list[pdopt.data.Constraint]): 
+            List containing the constraint objects.
+        model (pdopt.data.Model):
+            Model object containing the evaluation function.
+        set_id (int):
+            id of the set which the problem is defined on.
+        x_mask (list[float]):
+            Mask to filter out the discrete paramters, as these are
+            constant in the set and not modified by the optimiser.
+        l (list[float]):
+            Lower bounds of the continous input parameters.
+        u (list[float]):
+            Upper bounds of the continous input parameters.
+    '''     
+    
     def __init__(self, model, design_space, set_id, debug=False, **kwargs):
+        """
+        Initialise the NeuralNetwork Surrogate optimisation problem object.
+
+        Args:
+            model (pdopt.data.Model):
+                Model object containing the evaluation function.
+            design_space (pdopt.data.DesignSpace): 
+                The design space object of the problem.
+            set_id (int):
+                id of the set which the problem is defined on.
+            debug (bool, optional): 
+                Fix the random generator seed for testing purposes. Defaults to False.
+
+        Returns:
+            None.
+
+        """
+        
         self.model = model  # store reference to model
         self.design_space = design_space
 
@@ -123,6 +169,18 @@ class NNSurrogate(Problem):
         # Test, report if it has not been effective
 
     def _evaluate(self, X, out, *args, **kwargs):
+        """
+        Evaluation function for the pymoo optimisation algorithms using the Neural Network Surrogate
+
+        Args:
+            X (numpy.ndarray): Input values to be evaluated.
+            out (dict[str, numpy.ndarray]): A dictionary containing the evaluated objectives and constraint function.
+
+        Returns:
+            None.
+
+        """
+        
         X_in = []
         for x in X:
             in_x = []
@@ -174,6 +232,17 @@ class NNSurrogate(Problem):
         out["G"] = np.array(G_list)
 
     def recover_pts(self, X):
+        """
+        Reconstruct the true output from the evaluation function of the optimal points X.
+
+        Args:
+            X (numpy.ndarray): Input values to be evaluated.
+
+        Returns:
+            numpy.ndarray: Array containing the input values with the true objective and constraint values.
+
+        """
+        
         # Reconstruct the full input with the discrete variables
         X_in = []
 
@@ -220,6 +289,14 @@ class NNSurrogate(Problem):
         return self.recovered_out
 
     def sample(self):
+        """
+        Sample to train the surrogate model. 
+
+        Returns:
[…]
