+++ src/pdopt/exploration.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Created on Fri Oct 15 15:42:04 2021
+This module contains the functions and objects required to carry out the exploration analysis.
 
 """
 
@@ -46,7 +46,7 @@ __status__ = "Development"
 
 @contextlib.contextmanager
 def tqdm_joblib(tqdm_object):
-    """Context manager to patch joblib to report into tqdm progress bar given as argument"""
+    #Context manager to patch joblib to report into tqdm progress bar given as argument
 
     class TqdmBatchCompletionCallback(joblib.parallel.BatchCompletionCallBack):
         def __call__(self, *args, **kwargs):
@@ -62,9 +62,29 @@ def tqdm_joblib(tqdm_object):
         tqdm_object.close()
 
 
-def generate_surrogate_training_data(
-    parameters_list, model, n_train_points, save_dir=None, debug=False
-):
+def generate_surrogate_training_data(parameters_list, model, n_train_points, save_dir=None, debug=False):
+    """
+    Generates the training data from the evaluation function  for 
+    the probabilistic surrogate model, by sampling in the entire design space 
+    using a Sobol sequence.
+
+    Args:
+        parameters_list (list[pdopt.Parameter]): 
+            List containing the input parameter objects.
+        model (pdopt.data.Model): 
+            Model object containing the evaluation function.
+        n_train_points (int): 
+            Number of training datapoints to generate.
+        save_dir (str, optional): 
+            Filename where to save the generated data into a .csv file. If set to None, no file is generated.
+        debug (bool, optional): Fix random generator for testing purposes. Defaults to False.
+
+    Returns:
+        training_data (pandas.DataFrame): Dataframe with the generated datapoints, 
+        with input and output columns as defined in the PDOPT case.
+
+    """
+    
     # If factorial sampling has too many points, we switch to a LHS to save time
 
     # levels_list = [parameter.ranges for parameter in parameters_list]
@@ -136,9 +156,30 @@ def generate_surrogate_training_data(
     return training_data
 
 
-def generate_surrogate_test_data(
-    n_points, parameters_list, model, save_dir=None, debug=False
-):
+def generate_surrogate_test_data(n_points, parameters_list, model, save_dir=None, debug=False):
+    """
+    Generates the testing data from the evaluation function for 
+    the probabilistic surrogate model, by sampling in the entire design space 
+    using latin hybercube sampling.
+
+    Args:
+        n_points (int): 
+            Number of training datapoints to generate.
+        parameters_list (list[pdopt.Parameter]): 
+            List containing the input parameter objects.
+        model (pdopt.data.Model): 
+            Model object containing the evaluation function.
+        save_dir (str, optional): 
+            Filename where to save the generated data into a .csv file. If set to None, no file is generated.
+        debug (bool, optional): 
+            Fix random generator for testing purposes. Defaults to False.
+
+    Returns:
+        test_data (pandas.DataFrame): Dataframe with the generated datapoints, 
+        with input and output columns as defined in the PDOPT case.
+
+    """
+    
     # Generate random samples via Latin Hypercube and evaluate them
     if debug:
         samples = LatinHypercube(len(parameters_list), seed=42).random(n_points)
@@ -193,12 +234,31 @@ def generate_surrogate_test_data(
 
 def generate_input_samples(n_points, parameters_list, rule="lhs", debug=False):
     """
-    Random samples via Latin Hypercube and evaluate them
-    Rules:
-        - 'lhs'   : Latin Hypercube
-        - 'sobol' : Sobol Rules
-        - 'grid'  : Grid Hypercube
+    Auxiliary function for sampling the full design space.
[…]
