+++ docs/data.md
@@ -1,6 +1,27 @@
 # Documentation for `pdopt.data`
 
+Module that contains all the data structures used within `PDOPT`.
 
-## ::: pdopt.data
+# Design Set objects
 
+## ::: pdopt.data.DesignSpace
 
+## ::: pdopt.data.DesignSet
+
+# Input Parameter objects
+
+## ::: pdopt.data.ContinousParameter
+
+## ::: pdopt.data.DiscreteParameter
+
+# Response objects
+
+## ::: pdopt.data.Objective
+
+## ::: pdopt.data.Constraint
+
+# Simulation model objects
+
+## ::: pdopt.data.Model
+
+## ::: pdopt.data.ExtendableModel
