+++ docs/optimization.md
@@ -1,19 +1,12 @@
-# Documentation for `pdopt.data`
+# Documentation for `pdopt.optimisation`
+
+Module containing the functions and objects used in the Search phase.
+
+## ::: pdopt.optimisation.Optimisation
 
 
-## ::: pdopt.data.DesignSpace
-	handler : python
-	options:
-		members:
-			- __init__
-			- from_csv
-			- from_pickle
-			- save_to_pickle
-			- get_exploration_results
-			- get_optimum_results
-			- set_discard_status
-			- save_exploration_results
-			- save_optimisation_results
-			- sample_from_set
-		show_source: false
-		
+## ::: pdopt.optimisation.NNSurrogate
+
+## ::: pdopt.optimisation.KrigingSurrogate
+
+## ::: pdopt.optimisation.DirectOpt

[…]
