+++ README.md
@@ -39,7 +39,9 @@ Use the `--user` option if not running in admin mode.
 
 ## Documentation
 
-Framework documentation with example usage, case setup tutorial and API reference can be found in the docs folder of the GitHub repository.
+Online API documentation with a (WIP) example usage can be found here: https://pdopt-code.readthedocs.io/en/latest/
+
+An outdated PDF document with some explaination of the theory behind PDOPT can be found in the docs folder.
 Additional information can be found in the following paper: ![Application of Probabilistic Set-Based Design Exploration on the Energy Management of a Hybrid-Electric Aircraft
 ](https://www.mdpi.com/2226-4310/9/3/147)
 
