+++ docs/tutorial.md
@@ -1,27 +1,333 @@
 
+# Example PDOPT Case
 
-StatusCode        : 200
-StatusDescription : OK
-Content           : # Alias eminus purpureum dabat hominem poenae in
-                    
-                    ## Talia at in perpetuos dicere cervos
-                    
-                    Lorem *markdownum caedibus ignis*, esse nempe saepe propago dedisse. Thybridis
-                    saturae, deae matrem; et dedit ...
-RawContent        : HTTP/1.1 200 OK
-                    Transfer-Encoding: chunked
-                    Connection: keep-alive
-                    access-control-allow-origin: *
-                    Content-Type: text/plain; charset=utf-8
-                    Date: Wed, 24 Jan 2024 19:06:51 GMT
-                    Server: nginx/1.14.2...
-Forms             : {}
-Headers           : {[Transfer-Encoding, chunked], [Connection, keep-alive], [access-control-allow-origin, *], [Content-Type, text/plain; charset=utf-8]...}
-Images            : {}
-InputFields       : {}
-Links             : {}
-ParsedHtml        : mshtml.HTMLDocumentClass
-RawContentLength  : 2201
+The files of the example described in this chapter can be found in the [GitHub repoistory](https://github.com/spinjet/pdopt-code/tree/main/example).
 
+## Problem Setup
+
+The problem is a design space exploration of hybrid-electric aircraft energy management strategies, where the goal is to minimise fuel consumption and NOx emissions under a maximum take-off mass constraint. After understanding the problem to analyse, the user should identify the input parameters of the model and the responses he desires to optimise or impose constraints on. For this example, the following input parameters have been identified:
+
+|    Name    |    Type    | Range | Description                                    |
+|:----------:|:----------:|:-----:|------------------------------------------------|
+|  climb_h0 | continuous |       | The DOH at the beginning of the climb segment  |
+|  climb_h1 | continuous |       | The DOH at the end of the climb segment        |
+| cruise_h0 | continuous |       | The DOH at the beginning of the cruise segment |
+| cruise_h1 | continuous |       | The DOH at the end of the cruise segment       |
+
+The user should decide if these parameters are continuous or discrete, i.e. describe a range of possible values or a set of discrete choices. In the first case, the user should also fix the boundaries of this range. This information will be used in defining the `input.csv` file and the arguments of the evaluation function.
+
+Then the user should list the responses they wish to analyse and define them as objectives or constraints. Special care should be given to objectives, considered "soft constraints" for the exploration phase. This is useful if the user wishes to understand which areas of the design space are guaranteed to provide at most (or least) a certain value. For instance, the user wishes to see which areas of the design space are more likely to have a minimum value of NOx emissions under a certain figure. For the example problem, the following requirements are identified:
+
+|  Name  |    Type    |    Operator    | Value | Description                                 |
+|:------:|:----------:|:--------------:|:-----:|---------------------------------------------|
+|   Mf   |  objective | min (minimise) |  nan  | Fuel Burnt during the mission               |
+| M_NOx  |  objective | min (minimise) |  nan  | NOx Emission during the mission             |
+|   TOM  | constraint | lt (less than) | 20000 | Take-off mass, constrainted to 20000 kg     |
+
+These response parameters will then define the output of the evaluation function and the `responses.csv` file.
+
[…]
