+++ docs/usage.md
@@ -1,0 +1,123 @@
+
+# User Guide of PDOPT
+
+
+## Structure of a PDOPT case
+
+A PDOPT analysis is composed of the following elements:
+
+- A design model, i.e. a computational simulation of the object under design that takes in design parameters and returns quantites of interest (QoI) quantifying its performance o behaviour. For instance, a simulation model describing the design of a wing would take as input parameters geometry and materials and return as QoI its lift-to-drag ratio, its maximum lift coefficient, its structural weight, and so on.
+- A set of input parameters (Continous or Discrete) which define the design space.
+- Quantities of Interest, which are divided into Constraints and Objectives:
+  - Constraints are feasibility requirements, i.e. the design must satisfy them to be acceptable.
+  - Objectives are desirability requirements, i.e. what is expected by the design to do, and possible the best as possible.
+  
+The framework provides object-oriented classes to define these items within PDOPT. 
+The user has first to identify which variables in the design model are parameters (and therefore define a design space), and which instead describe a behaviour of charateristic of the system that they wish to control (either to optimise or to constraint). Usually the design process consists in a set of performance figures of merit to optimise given specific constraints.  
+
+After the definition of the variables and the set-up of the script, the workflow consists in running the exploration phase with `pdopt.exploration.ProbabilisticExploration` and the search phase with `pdopt.optimisation.Optimisation`.
+
+
+## Definition of the Simulation/Design Model
+
+The framework is designed to assume the simulation model as a black box. Hence, `PDOPT` only expects the data is presented through the interface. Both the `ProbabilisticExploration()` and `Optimisation()` objects expect the simulation model to have the `.run()` method for running the evaluation function. The framework provides two approaches to connect the simulation model to the API: the class `data.Model()` and the `data.ExtendableModel()` abstract class. The first is a wrapper for the actual evaluation function. The second is a template for defining a custom model object, useful if the simulation model requires a setup or parameters to be changed at the object creation. Examples of their usage is shown here:
+
+```python
+# Example using the ExtendableModel abstract class
+class My_Model(ExtendableModel):
+	
+	def __init__(self, model_parameters):
+	self.model_parameters = model_parameters
+		# Do something with the parameters to setup the simulation
+	
+	def run(self, *args):
+		# Do something with args
+	
+# Example using the Model object
+def my_function(self, *args):
+	# Do something with args
+	
+my_model = data.Model(my_function)
+```
+
+The evaluation function must be crafted such as:
+
+- The input arguments are an ordered list following the same order of the parameters specified in the `DesignSpace` obect.
+- The output quantities must be returned through a python dictionary whose keywords are the constraint or objective name specified in the `DesignSpace` object.
+
+Here is an example where the parameters `A, B, C` are inputs and `X, Y` are outputs:
+
+```python
+def my_function(A, B, C):
+	# do something
+	return {'X' : X, 'Y' : Y}
+```
+
+It is possible to construct an evaluation function with flexible inputs by using the Python star argument (*args). This is useful especially if the same function is being used in different `PDOPT` analyses:
+
+```python
+# Example of usage of star argument 
+def my_flexible_function(*args):
+	# Unpack parameters depending on length
+	if len(args) == 3:
+		A, B, C = args
+	else:
+		A, B, C, D = args
+	
+	## do something
+``` 
+
+## Definition of the Input Parameters
+
[…]
