### 2024-02-12 spinjet/pdopt-code#1 @spinjet
@jbussemaker

Online documentation is complete and online.
https://pdopt-code.readthedocs.io/en/latest/

Last task is creating a release version and publishing to PyPI.
