### 2024-02-12 #6110 @spinjet
Dear @jbussemaker @kyleniemeyer
I have completed the online documentation, can be found here: https://pdopt-code.readthedocs.io/en/latest/
Currently it sits on the joss review branch.

There's only versioning and creating a PyPI submission left in the reviewer's checklist.
I reckon once these are completed the review process is complete and ready for publication?
