### 2024-02-12 #6110 @jbussemaker
@spinjet yes indeed! Don't forget to update the installation instructions too once you have a PyPI release ;)
