### 2024-02-12 spinjet/pdopt-code#4 @e-dub
Great changes.  The paper has really advanced.  In rereading it, I have the following questions, critique, etc.

## Paper, part II
- [x] "Figure Figure", "Figures Figure", "Fig. Figure"
- [x] From the text, it is not clear if you are dealing with discrete or continuous design variables or both.  Is either phase limited by design variable type?  Can you mention this.
- [ ] Please mention if there are any limitations to responses included in optimization? Convexity, unimodality, continuity, smooth, bifurcated, etc.? Or number of variables or constraints? Could islands of feasibility in the design space be handled?
- [ ] Is there a better way to show "traditional design" versus your proposed set-based method in Figure 1?  I do not know if this is accurate (or it is not clear to me)
- [ ] Great that you have included a plot of the architecture. Can this be done a bit cleaner?  The arrows are entering the process from all sides and therefore I find it hard to follow.
- [ ] Decision boundary or feasibility boundary?
- [ ] The difference between "exploration" and "search" is not clear defined in text.  Please define and differentiate in text.

- [ ] The reviewer checklist requires a statement of need in documentation. Could you please add this?  "A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?" I find this related to the questions I have above about the software's usage and limitations.

(please repoen this issue)
