### 2024-02-13 spinjet/pdopt-code#4 @kyleniemeyer
@e-dub @spinjet regarding statement of need in the documentation: unlike in the paper, where we have the required section "Statement of Need", the docs don't have to explicitly have a header titled like that, but there does need to be a clear statement/description of need as described above. The README / docs landing page is often the best place for this.
