+++ paper/paper.bib
@@ -308,3 +308,13 @@ author = {Raul Yondo and Esther Andrés and Eusebio Valero},
   howpublished = {\url{http://github.com/SheffieldML/GPy}},
   year = {since 2012}
 }
+
+@article{UNSGA3,
+  author={Seada, Haitham and Deb, Kalyanmoy},
+  journal={IEEE Transactions on Evolutionary Computation}, 
+  title={A Unified Evolutionary Optimization Procedure for Single, Multiple, and Many Objectives}, 
+  year={2016},
+  volume={20},
+  number={3},
+  pages={358-369},
+}
