### 2024-02-13 spinjet/pdopt-code#4 @spinjet
Thank you for your feedback @e-dub.  I'll address the changes here:

> - "Figure Figure", "Figures Figure", "Fig. Figure"

This has been corrected.

> - From the text, it is not clear if you are dealing with discrete or continuous design variables or both.  Is either phase limited by design variable type?  Can you mention this.

The framework is designed to handle both types, it seems I forgot to explicitly mention this. It has been added in the first paragraph of the Statement of Need section.

> -  Please mention if there are any limitations to responses included in optimization? Convexity, unimodality, continuity, smooth, bifurcated, etc.? Or number of variables or constraints? Could islands of feasibility in the design space be handled?

Feasibility islands are intrinsically handled as the design space is divided into small portions, each one evaluated with the probabilistic exploration phase, and the ones with a sufficiently high probability of having feasible designs are then analysed with the optimiser (search phase).
Regarding the behavior of the evaluation function for optimisation, I've chosen to use UNSGA-3 because it's a gradient-free population based algorithm... which should handle even non-smooth functions (wich in case of multi-disciplinary models, is a possibility).
Finally, the code has not been tested with large number of objectives and input paramters yet (so far maximum 8 inputs, 3 objectives and 3 constraints), but its limitations are ultimately the computational cost of the model. The surrogate option for the search phase was added to enable handling large number of parameters as the GA would require a lot of functional evaluations in that case. These consideraitons have been now integrated in the 8th paragraph of the Statement of Need section, where the search phase is described.

> -  Is there a better way to show "traditional design" versus your proposed set-based method in Figure 1?  I do not know if this is accurate (or it is not clear to me)

I re-read the paragraph where the Figure is referenced and it was not sufficiently described there. The distinction between SBD and the classical design approach is in focusing on eliminating unfeasible designs and leaving options open rather than selecting a single "optimal" design point to refine iteratively. SBD is more robust because it allows more margin of movement in case adjustments on the designs are needed (the learning points in the timeline). I've expanded the paragraph to include these aspects, hopefully this makes the image more clear.

> -  Great that you have included a plot of the architecture. Can this be done a bit cleaner?  The arrows are entering the process from all sides and therefore I find it hard to follow.

The diagram is laid out such that at the centre the main process of the code is shown, with the inputs on the sides. Unfortunatelty this was the best layout I could come with as the image was from a presentation. I could try to emphasize the start and finish and what is the actual architecture and what is inputs with background boxeds.

> -  Decision boundary or feasibility boundary?

The term "decision" boundary was used because it's the edge where the set-based elimination process decides if one set is to be kept or not... this is because you may define a desirability constraint from the objectives. These act to restrict the design space to where a minimum level of improvement can be achieved (e.g. minimise F but you want F to be at least less than 100). These are not compulsory and generally useful after a run of PDOPT with just feasibility constraints. Hence why they are named "desirability", since these do not automatically mean the design should be rejected.

> -  The difference between "exploration" and "search" is not clear defined in text.  Please define and differentiate in text.

This is briefly mentioned in the second paragraph of the Statement of Need section. I've expanded it to be more clear, but the following paragraphs explaining how the exploration phase and search phase work should already indicate the difference between the two steps.

> # Documentation
> -  The reviewer checklist requires a statement of need in documentation. Could you please add this?  "A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?" I find this related to the questions I have above about the software's usage and limitations.

As mentioned by @kyleniemeyer, there is no need for a direct section but I have extended the description in the introduction page of the online documentation.

I've pushed the changes for the paper, please let me know if you need anything else.
