### 2024-02-13 #6110 @spinjet
https://github.com/spinjet/pdopt-code/issues/4#issuecomment-1942157939

Paper updated with the comments of the second reviewer.
