+++ README.md
@@ -52,3 +52,7 @@ This software is currently being maintained by me @spin-jet. If you find any bug
 
 This software was developed within Project ![FutPrInt50](https://futprint50.eu/), with EU Horizon 2020 Grant No. 875551.
 The authors wants to thank all the researchers in the project who cotnributed with their input to shape the framework.
+
+## Licensing
+
+Copyright (c) 2021 Cranfield University. This software is released under the permissive MIT License.
