### 2024-02-14 54177e9
Added release, updated installation method
