### 2024-02-14 spinjet/pdopt-code#1 @spinjet
Code has been published on PyPI: https://pypi.org/project/pdopt/
And I've made an official release as version 0.5.0.

@jbussemaker  Let me know if you are satisfied with the completion of the tasks.
