### 2024-02-14 spinjet/pdopt-code#1 @jbussemaker
@spinjet after these small updates, I am satisfied:
- [x] Link to the documentation website in the repository (click on the settings cog next to "About" and add the URL)
- [x] Fix small typos in the main readme
  - PyPI is styled exactly like that: "PyPI"
  - "**Alternatevly** it is possible to **donwload** the release..."
- [x] Also update installation instructions on the documentation website
- [x] On the documentation main page there is a typo in the acknowledgements: "cotnributed"
