+++ README.md
@@ -28,8 +28,13 @@ The optional visualisation tool requires the `dash-plotly` package to be install
 
 ## Installation
 
-First `git clone` the repository or download the zip file.
-Then run the setup from the root folder:
+The reccomended method is through PyPI by running the command:
+
+```
+pip install pdopt
+```
+
+Alternatevly  it is possible to donwload the release, or `git clone` the repository. Then run the setup from the root folder:
 
 ```
 pip install -e .
@@ -53,7 +58,7 @@ This software is currently being maintained by me @spin-jet. If you find any bug
 ## Acknowledgements
 
 This software was developed within Project ![FutPrInt50](https://futprint50.eu/), with EU Horizon 2020 Grant No. 875551.
-The authors wants to thank all the researchers in the project who cotnributed with their input to shape the framework.
+The authors wants to thank all the researchers in the project who contributed with their input to shape the framework.
 
 ## Licensing
 
