### 2024-02-15 spinjet/pdopt-code#1 @jbussemaker
Thank you! Then I will go ahead and close this issue, and recommend publication :)

Very nice work!
