### 2024-02-15 #6110 @jbussemaker
@kyleniemeyer all items in my checklist and issues have been addressed, I support publication of the JOSS paper :+1:
