### 2024-02-15 #6110 @kyleniemeyer
Thanks @jbussemaker!

@e-dub, do the changes made in https://github.com/spinjet/pdopt-code/issues/4 address your comments?
