### 2024-02-29 #6110 @e-dub
@kyleniemeyer my review checklist is complete and issues in my issue have been addressed. I support publication of the JOSS article.
