### 2024-02-29 #6110 @kyleniemeyer
@spinjet I'll add that at this point, the software will not change, so you can merge the reviewed branch (if that applies). The paper may change, but doesn't need to be in the final archive for the software—that can remain in a separate paper-only branch if desired.
