### 2024-03-01 #6110 @spinjet
[ ] Double check authors and affiliations (including ORCIDs)
[ ] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
[ ] Archive the release on Zenodo/figshare/etc and post the DOI here.
[ ] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
[ ] Make sure that the license listed for the archive is the same as the software license.
[ ] Merge the JOSS review branch into main, create a paper-only branch for tidyness.

Hi @kyleniemeyer,
I am glad the submission has been accepted, I will go through my checklist as soon as possible.
