### 2024-03-01 514c95d
Removed paper folder from main
