+++ paper/paper.md
@@ -1,102 +1,0 @@
-
----
-title: 'PDOPT: A Python library for Probabilistic Design space exploration and OPTimisation.'
-tags:
-  - Python
-  - Computational Engineering
-  - Design Space Exploration
-  - Set-Based Design
-  - Design Uncertainty
-  - Multi-Objective Optimization
-authors:
-  - name: Andrea Spinelli
-    orcid: 0000-0002-3064-3387
-    corresponding: true
-    affiliation: 1
-  - name: Timoleon Kipouros
-    orcid: 0000-0003-3392-283X
-    affiliation: 1
-affiliations:
- - name: Centre for Propulsion and Thermal Power, Cranfield University, MK430AL, UK
-   index: 1
-date: 22 August 2023
-bibliography: paper.bib
----
-
-# Summary
-
-Contemporary engineering design is characterised by products and systems with increasing complexity coupled with tighter requirements and tolerances. This leads to high epistemic uncertainty due to numerous possible configurations and a high number of design parameters. Set-Based Design is a methodology capable of handling these design problems, by exploring and evaluating as many alternatives as possible, before committing to a specific solution. 
-
-The Python package `PDOPT` aims to provide this capability without the high computational cost associated with the factorial-based design of experiments methods. Additionally, `PDOPT` performs the requirement mapping without explicit rule definition. Instead, it utilizes a probabilistic machine learning model to identify the areas of the design space most promising for user-provided requirements. This yields a plethora of feasible design points, assisting designers in understanding the system behaviour and selecting the desired configurations for further development.
-
-# State of the field
-
-Engineering design is a process where needs, quantitative and qualitative, are transformed into a set of specifications (e.g. geometry, materials, component list, etc) such that they can be sufficiently satisfied. The procedure involves performing several analyses of the system under design through suitable modelling tools, which map the design parameters to the quantified needs, and carrying out decisions to narrow and identify the range of desirable parameters. To aid this procedure, the designer relies on computer simulations through optimisation methods or the design of experiments in the design space.
-
-Historically, optimisation has been used in the late stages of design, to refine a solution already in the ballpark of the requirements. However, with the development of Multi-disciplinary Optimisation (MDO), there has been increasing interest in bringing optimisation in earlier stages of the design process. Searching for optimal designs with concurrent analysis of different disciplines allows one to obtain better results than with sequential optimisation alone, reducing the time and cost of the development cycle [@MartinsLambe2013].
-In the conceptual phase, optimisation can be employed to understand which combination of input design parameters enables to satisfy the requirements. The algorithm acts by changing the input quantities iteratively to find the ideal combination(s) that minimise/maximise quantities of interest (often performance targets) without violating the problem constraints (which can be technological, geometrical, economical, and so on). Population-based solvers such as Genetic Algorithm [@DebNSGA:2002] or Particle Swarm [@Kennedy1995] are best suited for this task as they evolve multiple design points in parallel, rather than iteratively improving a single solution [@Deb2008]. The result is a family of points, covering the design space which are optimal. Surrogate modelling [@Forrester2009] is often introduced to reduce the computational cost of each functional evaluation and correct with high-fidelity data fast low-fidelity models often used in the conceptual design phase [@Kontogiannis2020].
-
-Software libraries developed for optimisation in Python are the OpenMDAO framework, which focuses on multi-disciplinary optimisation [@Gray2019a], the Object-Oriented `pyOpt` [@Pyopt], the `DEAP` evolutionary optimisation library [@DEAP], and `pymoo` multi-objective optimisation library [@pymoo]. These libraries are open-source and general-purpose.
-
[…]
