+++ paper/paper.bib
@@ -1,320 +1,0 @@
-@misc{pdoptrepo,
-  author = {Andrea Spinelli},
-  title = {PDOPT},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub Repository},
-  url = {https://github.com/spinjet/pdopt-code}
-}
-
-@misc{fp50,
-  author = {},
-  title = {FutPrInt50},
-  year = {2020},
-  publisher = {},
-  journal = {},
-  url = {https://www.futprint50.eu/}
-}
-
-@article{SpinelliEASN:2021,
-doi = {10.1088/1757-899X/1226/1/012064},
-year = {2022},
-month = {feb},
-publisher = {IOP Publishing},
-volume = {1226},
-number = {1},
-pages = {012064},
-author = {Andrea Spinelli and Luchien Anderson and Hossein Balaghi Enalou and Bahareh Zaghari and Timoleon Kipouros and Panagiotis Laskaridis},
-title = {Application of Probabilistic principles to Set-Based Design for the optimisation of a hybrid-electric propulsion system},
-journal = {IOP Conference Series: Materials Science and Engineering},
-}
-
-@article{SpinelliEASN:2021,
-doi = {10.1088/1757-899X/1226/1/012064},
-year = {2022},
-month = {feb},
-publisher = {IOP Publishing},
-volume = {1226},
-number = {1},
-pages = {012064},
-author = {Andrea Spinelli and Luchien Anderson and Hossein Balaghi Enalou and Bahareh Zaghari and Timoleon Kipouros and Panagiotis Laskaridis},
-title = {Application of Probabilistic principles to Set-Based Design for the optimisation of a hybrid-electric propulsion system},
-journal = {IOP Conference Series: Materials Science and Engineering},
-}
-
-
-@article{SpinelliMDPI:2022,
-AUTHOR = {Spinelli, Andrea and Enalou, Hossein Balaghi and Zaghari, Bahareh and Kipouros, Timoleon and Laskaridis, Panagiotis},
-TITLE = {Application of Probabilistic Set-Based Design Exploration on the Energy Management of a Hybrid-Electric Aircraft},
-JOURNAL = {Aerospace},
-VOLUME = {9},
-YEAR = {2022},
-NUMBER = {3},
-ARTICLE-NUMBER = {147},
-ISSN = {2226-4310},
-DOI = {10.3390/aerospace9030147}
-}
-
-@article{SpinelliEASN:2022,
-doi = {10.1088/1742-6596/2526/1/012021},
-year = {2023},
-month = {jun},
-publisher = {IOP Publishing},
-volume = {2526},
-number = {1},
-pages = {012021},
-author = {Andrea Spinelli and Gustavo Pedro Krupa and Timoleon Kipouros and Boris Berseneff and Sébastien Fiette},
-title = {Investigation of the operational flexibility of a regional hybrid-electric aircraft},
-journal = {Journal of Physics: Conference Series},
-}
-
-
-@article{SpinelliAIAA:2023,
-author = {Andrea Spinelli and Gustavo P. Krupa and Timoleon Kipouros},
-title = {Set-Based Design Space Exploration to Investigate the Effect of Energy Storage Durability on the Energy Management Strategy of a Hybrid-Electric Aircraft},
-journal = {AIAA SCITECH 2023 Forum},
-year = {2023},
-month = {jan},
-chapter = {},
-pages = {},
-doi = {10.2514/6.2023-0837},
-}
-
-@article{MartinsLambe2013,
-author = {Martins, Joaquim R. R. A. and Lambe, Andrew B.},
-title = {Multidisciplinary Design Optimization: A Survey of Architectures},
-journal = {AIAA Journal},
-volume = {51},
-number = {9},
-pages = {2049-2075},
[…]
