### 2024-03-01 #6110 @spinjet
Dear @kyleniemeyer,

I've completed the checklist.

The released version for the JOSS publication is the latest one (v0.5.1) https://github.com/spinjet/pdopt-code/releases/tag/0.5.1
This has been archived on Zenodo: https://zenodo.org/doi/10.5281/zenodo.10732017

I hope this satisfies every requirement for publication.
