### 2024-03-01 76a1958
Updates JOSS paper
Minor typography edits
