### 2024-03-01 #6110 @kyleniemeyer
Hi @spinjet, just a few things needed:

- Please review and merge the changes I just made: https://github.com/spinjet/pdopt-code/pull/8
- Please correct the version in the Zenodo metadata (says v1 right now)
