### 2024-03-04 #6110 @spinjet
> Hi @spinjet, just a few things needed:
>
> * [*] Please review and merge the changes I just made: [Updates JOSS paper spinjet/pdopt-code#8](https://github.com/spinjet/pdopt-code/pull/8)
> * [*] Please correct the version in the Zenodo metadata (says v1 right now)

Hi @kyleniemeyer,
I've merged and corrected the Zenodo metadata.
