### 2024-03-04 #6110 @kyleniemeyer
Hi @spinjet, a few final edits in the references—sorry I didn't catch these before.

Please see the DOI errors above. Make sure to remove the `https://doi.org/` prefix in the `doi` field where appropriate, and check those two potential missing DOIs.
