### 2024-03-04 280724d
Update paper.bib
Fixed missing DOIs and formatting
