+++ paper/paper.bib
@@ -98,7 +98,7 @@ doi = {10.2514/1.J051895},
         year = {2024},
         volume = {188},
         pages = {103571},
-        doi = {https://doi.org/10.1016/j.advengsoft.2023.103571}
+        doi = {10.1016/j.advengsoft.2023.103571}
 		}
 
 @article{pymoo,
@@ -109,6 +109,7 @@ doi = {10.2514/1.J051895},
     volume={8},
     number={},
     pages={89497-89509},
+	doi={10.1109/ACCESS.2020.2990567}
 }
 
 @article{Gray2019a,
@@ -158,7 +159,7 @@ number = {1},
 pages = {50-79},
 year = {2009},
 issn = {0376-0421},
-doi = {https://doi.org/10.1016/j.paerosci.2008.11.001},
+doi = {10.1016/j.paerosci.2008.11.001},
 author = {Alexander I.J. Forrester and Andy J. Keane},
 }
 
@@ -179,7 +180,7 @@ volume = {97},
 pages = {105592},
 year = {2020},
 issn = {1270-9638},
-doi = {https://doi.org/10.1016/j.ast.2019.105592},
+doi = {10.1016/j.ast.2019.105592},
 author = {Spyridon G. Kontogiannis and Jean Demange and A. Mark Savill and Timoleon Kipouros},
 }
 
@@ -202,7 +203,7 @@ author = {Spyridon G. Kontogiannis and Jean Demange and A. Mark Savill and Timol
   volume  = {13},
   number  = {70},
   pages   = {2171-2175},
-  doi     = {https://doi.org/10.5555/2503308.2503311}
+  doi     = {10.5555/2503308.2503311}
 }
 
 @book{MontgomeryDE,
@@ -221,7 +222,7 @@ volume = {96},
 pages = {23-61},
 year = {2018},
 issn = {0376-0421},
-doi = {https://doi.org/10.1016/j.paerosci.2017.11.003},
+doi = {10.1016/j.paerosci.2017.11.003},
 author = {Raul Yondo and Esther Andrés and Eusebio Valero},
 }
 
@@ -317,4 +318,5 @@ author = {Raul Yondo and Esther Andrés and Eusebio Valero},
   volume={20},
   number={3},
   pages={358-369},
+  doi={10.1109/TEVC.2015.2459718}
 }
