+++ paper/paper.bib
@@ -202,8 +202,7 @@ author = {Spyridon G. Kontogiannis and Jean Demange and A. Mark Savill and Timol
   year    = {2012},
   volume  = {13},
   number  = {70},
-  pages   = {2171-2175},
-  doi     = {10.5555/2503308.2503311}
+  pages   = {2171-2175}
 }
 
 @book{MontgomeryDE,
@@ -300,7 +299,8 @@ author = {Raul Yondo and Esther Andrés and Eusebio Valero},
   volume={12},
   number={Oct},
   pages={2825--2830},
-  year={2011}
+  year={2011},
+  doi={10.48550/arXiv.1201.0490},
 }
 
 @Misc{gpy2014,
