### 2024-03-04 #6110 @kyleniemeyer
@spinjet it looks like there might be either a duplicate bibliography cite key or entry?
