+++ paper/paper.bib
@@ -16,18 +16,6 @@
   url = {https://www.futprint50.eu/}
 }
 
-@article{SpinelliEASN:2021,
-doi = {10.1088/1757-899X/1226/1/012064},
-year = {2022},
-month = {feb},
-publisher = {IOP Publishing},
-volume = {1226},
-number = {1},
-pages = {012064},
-author = {Andrea Spinelli and Luchien Anderson and Hossein Balaghi Enalou and Bahareh Zaghari and Timoleon Kipouros and Panagiotis Laskaridis},
-title = {Application of Probabilistic principles to Set-Based Design for the optimisation of a hybrid-electric propulsion system},
-journal = {IOP Conference Series: Materials Science and Engineering},
-}
 
 @article{SpinelliEASN:2021,
 doi = {10.1088/1757-899X/1226/1/012064},
