### 2024-03-05 #6110 @spinjet
@kyleniemeyer Should be fixed now, there was apparently one duplicated entry in the bib file.
